% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_samples.R
\name{analyze_samples}
\alias{analyze_samples}
\title{Analyze mass spectrometry data}
\usage{
analyze_samples(x, g, col.names, info.vars = c("Accession", "Sequence",
  "Annotated.Sequence", "Descriptions", "Modifications",
  "Reporter.Quan.Result.ID"))
}
\arguments{
\item{x}{data frame with expression values for both samples}

\item{g}{vector of factor levels for samples}

\item{col.names}{vector of column names for output data frame}

\item{info.vars}{vector of column names containing metadata information.
These variables are collapsed if not unique.}
}
\description{
Analyze mass spectrometry data for expression differences between sample 
groups.
}

