% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/assessSurvTime.R
\name{assessSurvTime}
\alias{assessSurvTime}
\title{Assess survival time}
\usage{
assessSurvTime(T1, T2, status)
}
\arguments{
\item{T1}{time 1}

\item{T2}{time 2}

\item{status}{censoring status}
}
\value{
A list with elements
\item{Otime}{Observation time}
\item{Stime}{Censoring time}
\item{Etime}{Time to end of study}
\item{KFT}{Known Function Time}
\item{RevKM}{Kaplan-Meier Time}
}
\description{
Given two survival times and censoring status, assesses the survival times...[To Be Completed]
}
\details{
Details to be filled.
}

