% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplotSum.R
\name{barplotSum}
\alias{barplotSum}
\title{Annotated Barplot}
\usage{
barplotSum(tx, ttl = "")
}
\arguments{
\item{ttl}{the title of the boxplot}

\item{var}{the variable to be plotted (must be categorical)}
}
\description{
Function to provide a barplot from a table of categorical variable with labels wraped
}
\author{
Aline Talhouk
}

