% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_cut.R
\name{best_cut}
\alias{best_cut}
\title{Find the best cutpoint for a covariate}
\usage{
best_cut(f, d, n = c("b", "t", "qd", "qn"), AIC.range = 3,
  nround = 3, plot = TRUE, filename = NULL, nrow = NULL,
  ncol = NULL, title = "", ...)
}
\arguments{
\item{f}{formula object}

\item{d}{data frame}

\item{n}{number of groups to transform variable into. Options are "b" (two),
"t" (three), "qd" (four), and "qn" (five)}

\item{AIC.range}{If range of AIC is within \code{AIC.range} units, the
likelihood is too flat. We choose the best cutpoint using the alternative
method.}

\item{nround}{number of digits to round AIC and p-value on plots}

\item{plot}{logical; If \code{TRUE}, shows the survival curves for each
cutpoint in a facetted figure}

\item{filename}{file name for saving a png image of figure}

\item{nrow}{number of rows in facetted plot}

\item{ncol}{number of columns in facetted plot}

\item{title}{title for plot}

\item{...}{additional arguments for \code{plot}}
}
\value{
A list with the following elements
\item{cuts}{vector of cutpoints considered}
\item{fits}{A list of \code{coxph} objects run for each cutpoint}
\item{results}{A table showing the likelihood ratio test p-value,
log likelihood, and AIC for each cutpoint}
\item{opt.cut}{optimal cutpoint value}
\item{flat.lik}{If \code{TRUE}, the likelihood was too flat and the
alternative method was used}
Additionally, if \code{plot = TRUE}, the function also returns KM survival
curves for each possible cutpoint.
}
\description{
Repeatedly finds cutpoints for an explanatory variable in a univariable Cox
PH model. Also plots survival curves for each cutpoint.
}
\details{
Takes the cutpoint resulting in the lowest AIC. If the range of AIC values is
within \code{AIC.range} units, take the cutpoint that results in the two groups
having the most similar numbers of events and cases. The function can cut a
variable into anywhere from 2 to 5 groups.
}
\author{
Derek Chiu
}
