% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_annotated.R
\name{boxplotSum}
\alias{boxplotSum}
\title{Annotated Boxplot}
\usage{
boxplotSum(var, ttl = "", digit = 1)
}
\arguments{
\item{var}{the variable to be plotted (must be numerical)}

\item{ttl}{the title of the boxplot}

\item{digit}{the number of digits used for rounding (defaults to 1)}
}
\description{
Function to provide a boxplot annotated with a 5 point summary.
}
\author{
Aline Talhouk
}
