% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_cuts.R
\name{build_cuts}
\alias{build_cuts}
\title{Build cutpoint variables}
\usage{
build_cuts(x, n = c("b", "t"), var.prefix = NULL)
}
\arguments{
\item{x}{vector to split by certain cutpoints}

\item{n}{either "b" for binarization (2 groups) or "t" for trinarization (3 groups)}

\item{var.prefix}{variable name prefix}
}
\value{
A data frame of cutpoint variables built from a categorical biomarker. The
number of columns correspond to all the ways the biomarker could be cut into \code{n}
bins. Each column name starts with a "b" or "t" for "binarization" or "trinarization",
respectively, with the levels being compared separated by "v".
}
\description{
Transforms an ordinal variable into two or three groups for cutpoint analysis.
}
\examples{
set.seed(1108)
x <- sample(0:4, size = 1000, replace = TRUE)
head(build_cuts(x, n = "b"))
head(build_cuts(x, n = "t"))
head(build_cuts(x, n = "t", var.prefix = "PHGDH"))
}
\author{
Derek Chiu
}

