% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_coxph.R
\name{censor_date}
\alias{censor_date}
\title{Date of censoring}
\usage{
censor_date(date_dx, date_lfu, cutoff_yr)
}
\arguments{
\item{date_dx}{date of diagnosis}

\item{date_lfu}{date of last follow-up}

\item{cutoff_yr}{number of years to cutoff at}
}
\value{
date of censoring
}
\description{
Calculate date of censoring using date of diagnosis and last follow-up.
}
\details{
The cutoff date is \code{cutoff_yr} years after the date of diagnosis, on December
31st. If the cutoff date is earlier than the date of last follow-up, then the
cutoff date is the new date of censoring, otherwise use the date of last
follow-up.
}
\examples{
date_dx <- as.Date("2003-03-08")
date_lfu <- as.Date("2009-08-09")
censor_date(date_dx, date_lfu, 5)
censor_date(date_dx, date_lfu, 6)
}
\author{
Derek Chiu
}
