% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisqTests.R
\name{chisqTests}
\alias{chisqTests}
\title{Chi-Squared Tests for independence in contingency tables}
\usage{
chisqTests(x)
}
\arguments{
\item{x}{an object of class \code{CrossTable} containing the contingency table}
}
\value{
A table with method name, test statistic, degrees of freedom, and
p-value reported for each Chi-squared test.
}
\description{
The Pearson's Chi-Squared test and likelihood ratio (G test) of independence
are performed on the data matrix
}
\examples{
# Example from documentation of CrossTable
library(descr)
data(esoph, package = "datasets")
ct <- CrossTable(esoph$alcgp, esoph$agegp, expected = TRUE,
                 chisq = FALSE, prop.chisq = FALSE,
                 dnn = c("Alcohol consumption", "Tobacco consumption"))
chisqTests(ct)
                 
# Better example
set.seed(1108)
A <- rbinom(100, 3, 0.2)
B <- rbinom(100, 4, 0.8)
ct <- CrossTable(A, B)
chisqTests(ct)
}
\author{
Derek Chiu
}
\seealso{
\code{\link{CrossTable}}
}

