% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_format.R
\name{chr_to_date}
\alias{chr_to_date}
\title{Change character to date}
\usage{
chr_to_date(x, date.origin = EXCEL.ORIGIN)
}
\arguments{
\item{x}{a number that represents the number of days after \code{date.origin}}

\item{date.origin}{the date from which we count the number of days passed}
}
\value{
A date object, converted from a character string.
}
\description{
Change a numeric value stored as a character string to a date object.
}
\details{
The default \code{date.origin} used is the Excel 1900 version.
}
\examples{
chr_to_date("41041")
}
\seealso{
Other date formatting functions:
\code{\link[=cleanDate]{cleanDate()}},
\code{\link[=formatDate]{formatDate()}},
\code{\link[=getFormat]{getFormat()}},
\code{\link[=numericToDate]{numericToDate()}}
}
\author{
Derek Chiu
}
\concept{date formatting functions}
