% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanGenes.R
\name{cleanGenes}
\alias{cleanGenes}
\title{Clean the Reactome FI gene set table}
\usage{
cleanGenes(data, save = TRUE, file.path = "")
}
\arguments{
\item{data}{data object}

\item{save}{if \code{TRUE}, the cleaned object is saved as a tab-delimited file.}

\item{file.path}{The file path to save the object.}
}
\value{
A cleaned gene set table.
}
\description{
The Cytoscape Reactome FI Geneset Analysis is cleaned before Enrichment Analysis.
}
\details{
The following cleaning procedures are made: 1) The "<" needs to be removed from the
FDR column. 2) Duplicate the GeneSet column. 3) Add a column of ones. 4) Replace
unicode characters with ASCII plain text.
}
\author{
Derek Chiu
}

