% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compareDate.R
\name{compareDate}
\alias{compareDate}
\title{Compare two dates}
\usage{
compareDate(d1, d2, date.format = MM.DD.YYYY, units = "days",
  existing.missing.codes = NA, return.missing.code = NA)
}
\description{
Compare two dates
}

