% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/confusionMatrix.R
\name{confusionMatrix}
\alias{confusionMatrix}
\title{Confusion matrix summaries}
\usage{
confusionMatrix(x, y, seed = 20, num.boot = 1000, conf.level = 0.95,
  digits = 4)
}
\arguments{
\item{x}{a vector of predicted classes}

\item{y}{a vector of reference classes}

\item{seed}{random seed for bootstrapping}

\item{num.boot}{number of times to bootstrap. Defaults to 1000.}

\item{conf.level}{confidence level. Defaults to 95\%.}

\item{digits}{number of digits to round summaries to}
}
\value{
A confusion matrix for the predicted and reference classes. Then
the estimated statistics along with bootstrapped confidence intervals.
}
\description{
Calculates summaries from cross-tabulated reference and prediction labels
for a two-class variable.
}
\details{
Given two dichotomous variables summarized in a confusion matrix, this
function provides performance summaries. The accuracy, sensitivity,
specificity, positive predictive value (PPV), negative predictive value
(NPV), and the kappa statistic, along with their bootstrapped confidence
intervals are returned.
}
\author{
Aline Talhouk, Derek Chiu
}

