% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/confusionResultToHtmlTable.R
\name{confusionResultToHtmlTable}
\alias{confusionResultToHtmlTable}
\title{Confusion matrix results in HTML}
\usage{
confusionResultToHtmlTable(prediction, reference, ref.description,
  round.digits.p.value, num.boot.for.ci = NA, seed = 12)
}
\description{
Confusion matrix results in HTML
}

