% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_compute.R
\name{date_compute}
\alias{date_compute}
\alias{addToDate}
\alias{diffDate}
\alias{compareDate}
\title{Date computations}
\usage{
addToDate(org.date, delta, date.format = "MM.DD.YYYY",
  units = c("days", "weeks", "months", "years"),
  existing.missing.codes = NA, return.missing.code = NA, sep = "/")

diffDate(d1, d2, date.format = "MM.DD.YYYY", units = c("days", "weeks",
  "months", "years"), existing.missing.codes = NA,
  return.missing.code = NA, sep = "/")

compareDate(d1, d2, date.format = "MM.DD.YYYY",
  existing.missing.codes = NA, return.missing.code = NA, sep = "/")
}
\arguments{
\item{org.date}{original date}

\item{delta}{amount of time to add to \code{org.date}}

\item{date.format}{how to format the resulting date}

\item{units}{for \code{addToDate}, the unit of time for \code{delta}; for
\code{diffDate}, the unit of time for which to take the difference.
Defaults to "days".}

\item{existing.missing.codes}{missing dates}

\item{return.missing.code}{what to return if there is a missing input}

\item{sep}{date separator string}

\item{d1}{later date}

\item{d2}{earlier date}
}
\value{
\code{addDate} returns the new date after adding \code{delta} to
\code{org.date}.

\code{diffDate} returns the difference between two dates \code{d1 - d2} in the specified unit of time.

\code{compareDate} returns 1 if \code{d1 > d2}, -1 if \code{d1 < d2},
and 0 if \code{d1 == d2}.
}
\description{
\code{addToDate} adds a period of time to a date string and returns a new
date string. \code{diffDate} takes the difference between two dates and
returns in the specified unit of time. \code{compareDate} compares the
difference between two dates.
}
\details{
If \code{delta} is negative, then the returned date will be earlier than
\code{org.date}. The output date format will be the same as the input date
format.

The unit of time to add to or return in can be in \code{"days"},
\code{"weeks"}, \code{"months"}, or \code{"years"}. \code{compareDate} calls
\code{diffDate} and returns integer values specifying which date is earlier
(or if they are the same). \code{d1} should be later than \code{d2} so the
function returns nonnegative values.
}
\examples{
## Adding to a date
addToDate("2014/07/08", 10, date.format = "YYYY.MM.DD")
addToDate("2014-07-08", 10, date.format = "YYYY.MM.DD", sep = "-")
addToDate("2014/07/08", 10, date.format = "YYYY.MM.DD", units = "months")
addToDate("2014/07/08", -10, date.format = "YYYY.MM.DD", units = "years")

## Date differences
# Later date comes first, subtracts the second date
diffDate("2003/03/21", "1992/01/27", date.format = "YYYY.MM.DD")

# Otherwise negative
diffDate("1992/01/27", "2003/03/21", date.format = "YYYY.MM.DD")

# Different separator
diffDate("2003-03-21", "1992-01-27", date.format = "YYYY.MM.DD", sep = "-")

## Date comparisons
compareDate("01/22/1949", "04/13/1950", date.format = "MM.DD.YYYY")
compareDate("04/13/1950", "04/13/1950", date.format = "MM.DD.YYYY")
compareDate("04/13/1959", "04/13/1950", date.format = "MM.DD.YYYY")
compareDate("01-22-1949", "04-13-1950", date.format = "MM.DD.YYYY", sep =
"-")
}
\author{
Samuel Leung, Derek Chiu
}
\concept{date computation functions}
