% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_extremes.R
\name{date_extremes}
\alias{date_extremes}
\alias{maxDate}
\alias{minDate}
\alias{maxDateArray}
\alias{minDateArray}
\title{Latest and Earliest Dates}
\usage{
maxDate(dates, na.rm = TRUE)

minDate(dates, na.rm = TRUE)

maxDateArray(t.arr, date.format = "MM.DD.YYYY",
  existing.missing.codes = NA, return.missing.code = NA, sep = "/")

minDateArray(t.arr, date.format = "MM.DD.YYYY",
  existing.missing.codes = NA, return.missing.code = NA, sep = "/")
}
\arguments{
\item{dates}{vector of dates}

\item{na.rm}{logical. Should NAs be removed?}

\item{t.arr}{array of date strings}

\item{date.format}{format of the array of dates}

\item{existing.missing.codes}{missing dates}

\item{return.missing.code}{what to return if there is a missing input}

\item{sep}{date separator. Defaults to "/"}

\item{...}{additional arguments to \code{formatDate}}
}
\value{
The latest or earliest date from a vector of dates.
}
\description{
Finds the latest (max) and earliest (min) date given a vector of dates. The
functions ending in \code{Array} find the extremes for an array of dates as
opposed to a vector.
}
\details{
The input vector should have dates formatted as YYYY-MM-DD. If \code{na.rm}
is not set to the default (\code{TRUE}) and \code{dates} has \code{NA}
values, then the function will also return \code{NA}.
}
\examples{
## No NA
t1 <- as.Date(c("2015-03-01", "2015-02-15", "2015-05-01"))
minDate(t1)
maxDate(t1)

## With NA
t2 <- as.Date(c("2015-03-01", "2015-02-15", NA, "2014-05-01"))
maxDate(t2)
minDate(t2)
minDate(t2, na.rm = FALSE)

## Array of dates
many.dates <- c("03/21/1992", "04/21/2013", "10/10/2015")
maxDateArray(many.dates)
minDateArray(many.dates)

many.dates <- c("2009-03-01", "2010-01-12", "2015-01-11")
maxDateArray(many.dates, sep = "-")
minDateArray(many.dates, sep = "-")

ties.dates <- c("2009-03-01", "2010-01-12", "2010-01-12")
maxDateArray(ties.dates, sep = "-")
minDateArray(ties.dates, sep = "-")
}
\author{
Samuel Leung, Derek Chiu
}
