% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_matrix.R
\name{design_matrix}
\alias{design_matrix}
\title{Design Matrix with interactions}
\usage{
design_matrix(data, order = 2)
}
\arguments{
\item{data}{numeric data}

\item{order}{complexity of interactions. Second order interactions would
comprise of all pairs of variables, third order interactions include all
triplets of variables, and so on. Defaults to 2.}
}
\value{
A design matrix with desired order of interactions for all variables
}
\description{
A design matrix with desired order of interactions.
}
\examples{
set.seed(2021)
df <- data.frame(matrix(rnorm(40), ncol = 4))
design_matrix(df)
}
\author{
Derek Chiu
}
