% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diffDate.R
\name{diffDate}
\alias{diffDate}
\title{Difference between two dates}
\usage{
diffDate(d1, d2, date.format = MM.DD.YYYY, units = "days",
  existing.missing.codes = NA, return.missing.code = NA)
}
\description{
Difference between two dates
}

