% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/doBarplot.R
\name{doBarplot}
\alias{doBarplot}
\title{Generates barplot}
\usage{
doBarplot(input.d, data.description, biomarker.var.name, biomarker.name,
  biomarker.value.names, digits = 3)
}
\description{
Generates barplot
}

