% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/doBoxplotAmongSubtypes.R
\name{doBoxplotAmongSubtypes}
\alias{doBoxplotAmongSubtypes}
\title{Do a boxplot among subtypes}
\usage{
doBoxplotAmongSubtypes(input.d, data.description, biomarker.var.name,
  biomarker.name, subtype.var.name, subtype.name, pch = 4, jitter = 0.1,
  digits = 2, boot.mean, ...)
}
\description{
Do a boxplot among subtypes
}

