% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/doCheckPHMulti.R
\name{doCheckPHMulti}
\alias{doCheckPHMulti}
\title{Check the proportional hazards assumption in multivariable model}
\usage{
doCheckPHMulti(data, var.name, var.description)
}
\description{
Check the proportional hazards assumption in multivariable model
}

