% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/doCohortCharacteristics.R
\name{doCohortCharacteristics}
\alias{doCohortCharacteristics}
\title{Generate cohort characteristics}
\usage{
doCohortCharacteristics(input.d, marker.name, marker.description, var.names,
  is.var.continuous, var.descriptions, marker.value.labels.tolower = TRUE,
  show.missing = TRUE, show.missing.continuous = TRUE,
  do.droplevels = TRUE, show.percent = "both", stat.tests = NULL,
  stat.test.column.header = "association/correlation test",
  round.digits.p.value = 4, num.boot.for.ci = 1000,
  missing.codes.highlight = NULL, missing.codes = c("N/A", "", "Unk"),
  decimal = 0, caption = NA, html.table.border = 0, banded.rows = FALSE,
  css.class.name.odd = "odd", css.class.name.even = "even")
}
\arguments{
\item{input.d:}{The dataframe containing the data}

\item{marker.name:}{The variable that you want to split into different columns}

\item{var.names:}{The variables that you want the statistics for}

\item{is.var.continuous:}{Vector of length equal to the length of var.names with 1 indicating a continuous variable and 0 otherwise (this should be inferred in the function)}

\item{var.descriptions:}{Vector of strings to describe the variables as they are to appear in the table}
}
\description{
Generate cohort characteristics
}

