% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doCoxphGeneric.R, R/doCoxphMultivariable.R
\name{doCoxph}
\alias{doCoxph}
\alias{doCoxphGeneric}
\alias{doCoxphMultivariable}
\title{Fit a Cox proportional hazards regression model}
\usage{
doCoxphGeneric(
  input.d,
  var.names,
  var.descriptions,
  show.var.detail = FALSE,
  show.group.name.for.bin.var = FALSE,
  var.ref.groups = NULL,
  var.names.surv.time = c("os.yrs", "dss.yrs", "rfs.yrs"),
  var.names.surv.time2 = NULL,
  var.names.surv.status = c("os.sts", "dss.sts", "rfs.sts"),
  event.codes.surv = c("os.event", "dss.event", "rfs.event"),
  surv.descriptions = c("OS", "DSS", "PFS"),
  var.strata = NULL,
  missing.codes = c("N/A", "", "Unk"),
  use.firth = 1,
  firth.caption = FIRTH.CAPTION,
  add_log_hr = FALSE,
  stat.test = "waldtest",
  bold_pval = FALSE,
  sig.level = 0.05,
  round.digits.p.value = 4,
  round.small = FALSE,
  scientific = FALSE,
  caption = NA,
  html.table.border = 0,
  banded.rows = FALSE,
  css.class.name.odd = "odd",
  css.class.name.even = "even",
  split.table = 300,
  ...
)

doCoxphMultivariable(
  input.d,
  var.names,
  var.descriptions,
  show.var.detail = FALSE,
  show.group.name.for.bin.var = FALSE,
  var.ref.groups = NULL,
  var.names.surv.time = c("os.yrs", "dss.yrs", "rfs.yrs"),
  var.names.surv.time2 = NULL,
  var.names.surv.status = c("os.sts", "dss.sts", "rfs.sts"),
  event.codes.surv = c("os.event", "dss.event", "rfs.event"),
  surv.descriptions = c("OS", "DSS", "PFS"),
  var.strata = NULL,
  missing.codes = c("N/A", "", "Unk"),
  use.firth = 1,
  firth.caption = FIRTH.CAPTION,
  add_log_hr = FALSE,
  stat.test = "waldtest",
  bold_pval = FALSE,
  sig.level = 0.05,
  round.digits.p.value = 4,
  round.small = FALSE,
  scientific = FALSE,
  caption = NA,
  html.table.border = 0,
  banded.rows = FALSE,
  css.class.name.odd = "odd",
  css.class.name.even = "even",
  split.table = 300,
  format = c("long", "wide"),
  ...
)
}
\arguments{
\item{input.d}{The \code{data.frame} containing the data}

\item{var.names}{variables to include as predictors}

\item{var.descriptions}{vector of strings to describe the variables as they
are to appear in the table}

\item{show.var.detail}{logical. If \code{TRUE}, details such as categories and the
reference group for categorical variables are shown.}

\item{show.group.name.for.bin.var}{logical. If \code{TRUE}, the non-reference
group name is shown beside the hazard ratio for dichotomized variables.}

\item{var.ref.groups}{a vector of reference groups. If \code{NULL}, assume all
variables are binary/continuous. If an item in the vector is \code{NA}, assume
that particular marker is binary or continuous (i.e., treat it as a numeric
variable)}

\item{var.names.surv.time}{variable names of survival time}

\item{var.names.surv.time2}{optional variable names for interval ending
times, used when survival is left-truncated. The interpretation of
\code{var.names.surv.time} becomes interval starting times when this argument is
not \code{NULL}.}

\item{var.names.surv.status}{variable names of survival status}

\item{event.codes.surv}{event coding of survival status variable}

\item{surv.descriptions}{names abbreviated survival endpoints in returned
output}

\item{var.strata}{optional variable name to stratify by using
\code{\link[survival:strata]{survival::strata()}}}

\item{missing.codes}{character strings of missing values used in \code{input.d}}

\item{use.firth}{percentage of censored cases before using Firth's method for
Cox regression. If \code{use.firth = 1} (default), Firth is never used and if
\code{use.firth = -1} Firth is always used.}

\item{firth.caption}{subscript in html table output indicating Firth was used}

\item{add_log_hr}{if \code{TRUE}, show the log hazard ratio}

\item{stat.test}{the overall model test to perform on the Cox regression
model. Can be any of "waldtest", "logtest", or "sctest". If Firth is used,
only "logtest" can be performed. Test p-values are never Firth corrected (
per instruction from Aline 2015-04-14,15).}

\item{bold_pval}{logical; if \code{TRUE}, p-values are bolded if statistically
significant at \code{sig.level}}

\item{sig.level}{significance level; default 0.05}

\item{round.digits.p.value}{number of digits for p-value}

\item{round.small}{if \code{TRUE}, uses small number rounding via \code{\link[=round_small]{round_small()}}}

\item{scientific}{if \code{TRUE}, uses scientific notation when rounding}

\item{caption}{caption for returned object}

\item{html.table.border}{the border type to use for html tables}

\item{banded.rows}{logical. If \code{TRUE}, rows have alternating shading colour}

\item{css.class.name.odd}{Used to set the row colour for odd rows}

\item{css.class.name.even}{Used to set the row colour for even rows}

\item{split.table}{number of characters per row before splitting the table.
Applies to the pandoc table output.}

\item{...}{additional arguments to \code{\link[pander:pandoc.table.return]{pander::pandoc.table.return()}}}

\item{format}{Either a "long" or "wide" format for the \code{result.table.bamboo}
result element}
}
\value{
A list with the following elements
\itemize{
\item \code{result.table}: a data frame with the Cox model results.
\item \code{result.table.bamboo}: results transformed into a pandoc-ready format.
Display using \code{cat()}.
\item \code{result.table.html}: results transformed into an HTML-ready format. Display
using \code{\link[htmlTable:htmlTable]{htmlTable::htmlTable()}}.
\item \code{cox.stats}: present for \code{doCoxphMultivariable()} only. A list of
additional Cox model output statistics.
}
}
\description{
Fits \code{coxph} on all survival endpoints. \code{doCoxphGeneric} fits univariable
models for each variable in \code{var.names}, and \code{doCoxphMultivariable} fits a
multivariable model for all variables together in \code{var.names}.
}
\details{
Please note the following assumptions:
\itemize{
\item Marker can be binary, continuous or categorical.
\item Missing survival time/status variables are coded as \code{NA} (i.e. will only be
checked by \code{is.na()}).
\item Survival time/status variable name specified in the following order: "os",
"dss", "rfs".
\item Coding of survival status is binary only (i.e. cannot take survival status
of > 2 categories).
}
}
\examples{
library(survival)
doCoxphGeneric(input.d = lung, var.names = "sex", var.descriptions = "Sex",
               var.names.surv.time = "time",
               var.names.surv.status = "status", event.codes.surv = "2",
               surv.descriptions = "OS", caption = "")
}
\author{
Samuel Leung, Aline Talhouk, Derek Chiu
}
