% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/doCoxphContinuous.R
\name{doCoxphContinuous}
\alias{doCoxphContinuous}
\title{coxph model continuous}
\usage{
doCoxphContinuous(input.d, var.names, var.descriptions,
  var.names.surv.time = c("os.yrs", "dss.yrs", "rfs.yrs"),
  var.names.surv.status = c("os.sts", "dss.sts", "rfs.sts"),
  event.codes.surv = c("os.event", "dss.event", "rfs.event"),
  surv.descriptions = c("OS", "DSS", "RFS"), use.firth = 1,
  stat.test = "waldtest", round.digits.p.value = 4, caption = NA,
  html.table.border = 0, banded.rows = FALSE, css.class.name.odd = "odd",
  css.class.name.even = "even")
}
\description{
coxph model continuous
}

