% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/doCoxphMultivariable.R
\name{doCoxphMultivariable}
\alias{doCoxphMultivariable}
\title{coxph model multivariable}
\usage{
doCoxphMultivariable(input.d, var.names, var.descriptions,
  show.var.detail = FALSE, show.group.name.for.bin.var = FALSE,
  var.ref.groups = NULL, var.names.surv.time = c("os.yrs", "dss.yrs",
  "rfs.yrs"), var.names.surv.status = c("os.sts", "dss.sts", "rfs.sts"),
  event.codes.surv = c("os.event", "dss.event", "rfs.event"),
  surv.descriptions = c("OS", "DSS", "PFS"), missing.codes = c("N/A", "",
  "Unk"), use.firth = 1, firth.caption = FIRTH.CAPTION,
  stat.test = "waldtest", round.digits.p.value = 4, caption = NA,
  html.table.border = 0, banded.rows = FALSE, css.class.name.odd = "odd",
  css.class.name.even = "even", split.table = 300, ...)
}
\description{
coxph model multivariable
}

