% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/doGrdTrendBarplot.R
\name{doGrdTrendBarplot}
\alias{doGrdTrendBarplot}
\title{Grade trend barplot}
\usage{
doGrdTrendBarplot(input.d, title, legend.space = 0.2, use.prop = TRUE)
}
\arguments{
\item{input.d}{input \code{data.frame}}

\item{title}{title of barplot}

\item{legend.space}{amount of space between legend and edge of plot}

\item{use.prop}{logical; if \code{TRUE}, proportions are used}
}
\value{
a barplot showing the change of Grade over time
}
\description{
Grade trend barplot
}
\author{
Samuel Leung
}

