% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/doJitterplotAmongSubtypes.R
\name{doJitterplotAmongSubtypes}
\alias{doJitterplotAmongSubtypes}
\title{Do a jitterplot among subtypes}
\usage{
doJitterplotAmongSubtypes(input.d, data.description, biomarker.var.name,
  biomarker.name, subtype.var.name, subtype.name, pch = ".", jitter = 0.05,
  digits = 3, cex.axis = 0.9)
}
\description{
Do a jitterplot among subtypes
}

