% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/doKMPlots.R
\name{doKMPlots}
\alias{doKMPlots}
\title{Make Kaplan-Meier plots}
\usage{
doKMPlots(input.d, var.name, var.description, line.color = NULL,
  line.pattern = NULL, km.plot.ref.group = "single",
  single.test.type = "logrank", surv.type = "os",
  use.firth = FIRTH.THRESHOLD, use.aline.plot = FALSE, ...)
}
\description{
Make Kaplan-Meier plots
}

