% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/doKMPlotsPOLEInitTrt.R
\name{doKMPlotsPOLEInitTrt}
\alias{doKMPlotsPOLEInitTrt}
\title{Make Kaplan-Meier plots with initial treatment}
\usage{
doKMPlotsPOLEInitTrt(data, surv.type, use.aline.plot = FALSE, ...)
}
\description{
Make Kaplan-Meier plots with initial treatment
}

