% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/doSurvTrendBarplot.R
\name{doSurvTrendBarplot}
\alias{doSurvTrendBarplot}
\title{Survival trend barplot}
\usage{
doSurvTrendBarplot(input.d, title, surv.yrs, surv.sts, surv.event, event.desc,
  censor.desc, obs.yrs = 3, legend.space = 0.2)
}
\description{
Survival trend barplot
}

