% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/doTrtTrendBarplot.R
\name{doTrtTrendBarplot}
\alias{doTrtTrendBarplot}
\title{Treatment trend barplots}
\usage{
doTrtTrendBarplot(input.d, title, legend.space = 0.2, use.prop = TRUE)
}
\arguments{
\item{input.d}{input \code{data.frame}}

\item{title}{title of barplot}

\item{legend.space}{amount of space between legend and edge of plot}

\item{use.prop}{logical; if \code{TRUE}, proportions are used}
}
\value{
a barplot showing the change of Treatment over time
}
\description{
Treatment trend barplots
}
\author{
Samuel Leung
}

