% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_string.R
\name{escapeForGrep}
\alias{escapeForGrep}
\title{Escape string for regular expression}
\usage{
escapeForGrep(x)
}
\arguments{
\item{x}{a character vector}
}
\value{
A character vector with opening and closing square brackets and
parentheses escaped for use in \code{grep}.
}
\description{
Escape \code{[}, \code{]}, \code{(}, and \code{)} for use in \code{grep}.
}
\examples{
escapeForGrep("[index]")
escapeForGrep("(parentheses)")
}
\author{
Samuel Leung
}
