% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploreData.R
\name{exploreData}
\alias{exploreData}
\title{Graphical Exploration of a Dataset}
\usage{
exploreData(datmat)
}
\arguments{
\item{datmat}{the data frame (only categorical and numerical variables will
be analysed)}
}
\description{
Function to provide a graphical exploration of a dataset will print results
to file.
}
\author{
Aline Talhouk
}

