% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_results}
\alias{extract_results}
\alias{extract_km}
\alias{extract_cox}
\title{Extract results for executive summary}
\usage{
extract_km(data, var, outcome, args, digits = 3)

extract_cox(x, var, outcome)
}
\arguments{
\item{data}{data frame with clinical variables and outcomes}

\item{var}{variable name}

\item{outcome}{survival outcome. One of "os", "dss", "pfs".}

\item{args}{list of arguments for survival analyses}

\item{digits}{number of digits to round p-value}

\item{x}{an object from result of \code{doCoxphGeneric()} or \code{doCoxphMultivariable()}}
}
\value{
Extracted result that can be used in the form of inline text as part
of the Executive Summary.
}
\description{
Extract statistical test results for Executive Summary section at beginning
of an RSF.
}
\details{
\itemize{
\item \code{extract_km()} extracts the Kaplan-Meier Log Rank test p-value for a
specified variable
\item \code{extract_cox()} extracts Cox proportional hazards model hazard ratios with
confidence intervals and LRT p-values for a specified variable
}
}
\author{
Derek Chiu
}
