% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_format.R
\name{formatDate}
\alias{formatDate}
\title{Format a date.}
\usage{
formatDate(
  d,
  m,
  y,
  date.format = c("MM.DD.YYYY", "MMM.DD.YYYY", "DD.MM.YYYY", "DD.MMM.YYYY", "YYYY.MM.DD",
    "YYYY.MMM.DD"),
  sep = "/"
)
}
\arguments{
\item{d}{day of the month (1-31)}

\item{m}{month of the year (1-12)}

\item{y}{year of date}

\item{date.format}{how to format the date. Defaults to "month/day/year".}

\item{sep}{string used to separate \code{d}, \code{m}, and \code{y}. Defaults
to "/".}
}
\value{
A character string of a formatted date.
}
\description{
Prints a date into a pretty format.
}
\details{
Given the day, month, and year of a date, returns the date in a specific
format. The order and separating string can be modified using
\code{date.format} and \code{sep} respectively. Take note the order of the
arguments: day, month, and year. Only accepts "MM.DD.YYYY", "MMM.DD.YYYY",
"DD.MM.YYYY", "DD.MMM.YYYY", "YYYY.MM.DD", "YYYY.MMM.DD".
}
\examples{
formatDate(8, 7, 2011)
formatDate(8, 7, 2011, date.format = "YYYY.MM.DD")
formatDate(8, 7, 2011, date.format = "DD.MM.YYYY", sep = "-")
formatDate(10, 1, 2015, date.format = "MMM.DD.YYYY", sep = "-")
}
\seealso{
Other date formatting functions:
\code{\link[=chr_to_date]{chr_to_date()}},
\code{\link[=cleanDate]{cleanDate()}},
\code{\link[=getFormat]{getFormat()}},
\code{\link[=numericToDate]{numericToDate()}}
}
\author{
Samuel Leung, Derek Chiu
}
\concept{date formatting functions}
