% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatNA.R
\name{formatNA}
\alias{formatNA}
\title{Missing Value Formatting Function}
\usage{
formatNA(y, type = "cat")
}
\arguments{
\item{y}{a vector.}

\item{type}{whether the variable is "cat" or "cont". Defaults to cat.}
}
\value{
A categorical or numerical vector with all missing as formatted NA.
}
\description{
Takes a numeric vector and replaces all Missing codes with NA and returns a factor if the variable is categorical or a numeric variable if it's numeric.
}
\author{
Aline Talhouk
}

