% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{g_legend}
\alias{g_legend}
\title{Generate a legend}
\usage{
g_legend(a.gplot)
}
\arguments{
\item{a.gplot}{ggplot object}
}
\value{
ggplot object with legend
}
\description{
Given a ggplot object, generates a legend
}
\author{
Aline Talhouk
}
