% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_format.R
\name{getFormat}
\alias{getFormat}
\title{Get date format from character text}
\usage{
getFormat(
  date,
  char.format = c("MM.DD.YYYY", "MMM.DD.YYYY", "DD.MM.YYYY", "DD.MMM.YYYY", "YYYY.MM.DD",
    "YYYY.MMM.DD"),
  sep = ""
)
}
\arguments{
\item{date}{character string of date}

\item{char.format}{character text format of date}

\item{sep}{character string separating \code{date}}
}
\value{
A character string representing the POSIX standard date format
equivalent of the string in \code{char.format}.
}
\description{
Get the POSIX standard date formats from character text formats.
}
\examples{
getFormat("12/09/1993", "MM.DD.YYYY")
getFormat("2005-09-13", "YYYY.MM.DD")
}
\seealso{
Other date formatting functions:
\code{\link[=chr_to_date]{chr_to_date()}},
\code{\link[=cleanDate]{cleanDate()}},
\code{\link[=formatDate]{formatDate()}},
\code{\link[=numericToDate]{numericToDate()}}
}
\author{
Derek Chiu
}
\concept{date formatting functions}
