% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPval.R
\name{getPval}
\alias{getPval}
\title{Get the p-value}
\usage{
getPval(x)
}
\arguments{
\item{x}{a vector or matrix}
}
\value{
the Chi-squared p-value
}
\description{
Get the p-value
}
\references{
Robert Gentleman (http://tolstoy.newcastle.edu.au/R/help/01c/2809.html)
}

