% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_diagnostic_prev.R
\name{gg_diagnostic_prev}
\alias{gg_diagnostic_prev}
\title{Diagnostic Prevalence}
\usage{
gg_diagnostic_prev(data, se, sp, p, result = c("PPV", "NPV"))
}
\arguments{
\item{data}{input data with metrics}

\item{se}{column name for sensitivity}

\item{sp}{column name for specificity}

\item{p}{column name for prevalence}

\item{result}{whether to return the PPV or NPV}
}
\value{
A graph of the NPV or PPV by specificity, sensitivty, and prevalence.
}
\description{
Compare PPV or NPV by sensitivity, specificity, and prevalence.
}
\details{
A graph of positive predictive value (PPV) or negative predictive value (NPV)
is shown a function of specificity. Sensitivity is separated by line types,
and prevalence is shown as different facets.
}
\examples{
df <- data.frame(
  se = rep(c(0.5, 0.7, 0.9), each = 240),
  sp = rep(seq(0.21, 1, 0.01), 9),
  p = rep(c(0.32, 0.09, 0.026), length = 720)
)
gg_diagnostic_prev(df, se, sp, p, result = "PPV")
gg_diagnostic_prev(df, se, sp, p, result = "NPV")
}
\author{
Derek Chiu
}
