% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_annotated.R
\name{histSum}
\alias{histSum}
\title{Summary histogram}
\usage{
histSum(var, xlab = "", txt = "", sub = "", digit = 1)
}
\arguments{
\item{var}{a variable to be plotted}

\item{xlab}{X axis label}

\item{txt}{text to be used in the title (defaults to none)}

\item{sub}{a sublabel (defaults to none)}

\item{digit}{number of digits used in rounding (defaults to 1)}
}
\description{
Function to create an annotated histogram, with density plot
}
\author{
Aline Talhouk
}
