% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_string.R
\name{isFirstLetterUpperCase}
\alias{isFirstLetterUpperCase}
\title{Is the first letter uppercase?}
\usage{
isFirstLetterUpperCase(x)
}
\arguments{
\item{x}{character string}
}
\value{
logical; if \code{TRUE}, the first letter of the input string is
uppercase.
}
\description{
Returns a logical indicating whether the first letter of a character string
is uppercase.
}
\details{
If the input is an empty string, the function returns \code{TRUE}.
}
\examples{
isFirstLetterUpperCase("peanut butter")
isFirstLetterUpperCase("peanut Butter")
isFirstLetterUpperCase("Samuel butter")
isFirstLetterUpperCase("")
}
\author{
Samuel Leung, Derek Chiu
}
