% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/isUpper.R
\name{isUpper}
\alias{isUpper}
\title{Is the first letter uppercase?}
\usage{
isUpper(x)
}
\arguments{
\item{x}{character string}
}
\value{
logical; if \code{TRUE}, the first letter of the input
string is uppercase.
}
\description{
Returns a logical indicating whether the first letter
of a character string is uppercase.
}
\details{
If the input is an empty string, the function returns \code{TRUE}.
}
\examples{
isUpper("peanut butter")
isUpper("peanut Butter")
isUpper("Samuel butter")
isUpper("")
}
\author{
Samuel Leung
}

