% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kappaBootCI.R
\name{kappaBootCI}
\alias{kappaBootCI}
\title{Bootstrapped confidence interval for kappa statistic.}
\usage{
kappaBootCI(x, y, seed = 20, num.boot = 1000, conf.level = 0.95,
  method = c("cohen", "weighted", "fleiss", "krippendorff"),
  type = "nominal")
}
\arguments{
\item{x}{vector of binary scores from first rater}

\item{y}{vector of binary scores from second rater}

\item{seed}{random seed for bootstrapping}

\item{num.boot}{number of times to bootstrap. Defaults to 1000.}

\item{conf.level}{confidence level. Defaults to 95\%.}

\item{method}{statistic to calculate bootstrap estimate}

\item{type}{method setting for \code{method = "krippendorff"}}
}
\value{
bootstraped confidence interval for Cohen's kappa.
}
\description{
Function to compute cohen's kappa with binary data with a bootstrap
confidence interval
}
\details{
Cohen's kappa measures the amount of agreement between 2 raters of a binary
variable. The bootstrap confidence interval is adjusted (BCa).
}
\examples{
a <- rbinom(n = 100, size = 1, prob = 0.3)
b <- rbinom(n = 100, size = 1, prob = 0.7)
kappaBootCI(a, b)

## Use a different seed
kappaBootCI(a, b, 5)
}
\author{
Aline Talhouk, Derek Chiu
}
