% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_with_date.R
\name{knit_with_date}
\alias{knit_with_date}
\title{Knit R markdown report with today's date}
\usage{
knit_with_date(input, sep = "_", ...)
}
\arguments{
\item{input}{input file}

\item{sep}{separator between input file name and date}

\item{...}{additional arguments to \code{\link[rmarkdown:render]{rmarkdown::render()}}}
}
\value{
The output file name will have today's date appended at the end.
}
\description{
Append today's date to R markdown output file name. This function modifies
the default behaviour of the Knit button in R Studio. It is intended to be
passed in the YAML \verb{knit:} field.
}
\references{
\url{https://bookdown.org/yihui/rmarkdown-cookbook/custom-knit.html}
}
\author{
Derek Chiu
}
