% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifetable.R
\name{lifetable}
\alias{lifetable}
\title{Generate lifetables for multiclass variables}
\usage{
lifetable(obj, ntimes = 3, times = round(quantile(obj$time, 1/ntimes *
  rep(1:ntimes))), nround = 3, var.name = TRUE)
}
\arguments{
\item{obj}{An object of class \code{survfit}}

\item{ntimes}{number of time intervals}

\item{times}{A vector of endpoints of time intervals to show life table calculations.
By default, these are \code{ntimes} evenly spaced out endpoints based on the full range
of survival times.}

\item{nround}{number of digits to round table values}

\item{var.name}{logical; if \code{TRUE} (default), the variable name is appended to the
beginning of each stratum in the lifetable's \code{strata} column}
}
\value{
A table with the following columns:
\item{strata}{name of specific group in variable}
\item{times}{time interval}
\item{nsubs}{See \code{\link{lifetab}}}
\item{nlost}{See \code{\link{lifetab}}}
\item{nrisk}{See \code{\link{lifetab}}}
\item{nevent}{See \code{\link{lifetab}}}
\item{surv}{See \code{\link{lifetab}}}
\item{pdf}{See \code{\link{lifetab}}}
\item{hazard}{See \code{\link{lifetab}}}
\item{se.surv}{See \code{\link{lifetab}}}
\item{se.pdf}{See \code{\link{lifetab}}}
\item{se.hazard}{See \code{\link{lifetab}}}
}
\description{
Specify vector of time endpoints and create a cohort life table for two or more strata
}
\details{
Essentially a wrapper around \code{lifetab} that allows the user to input a \code{survfit}
object instead of vectors of raw values.
}
\examples{
obj <- survfit(Surv(futime, fustat) ~ rx, data = ovarian)
lifetable(obj)
lifetable(obj, ntimes = 4, var.name = FALSE)
lifetable(obj, ntimes = 4, times = c(200, 500, 800, 1000))
}
\author{
Derek Chiu
}
\seealso{
\code{\link{lifetab}}
}

