% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistfPrint.R
\name{logistfPrint}
\alias{logistfPrint}
\title{Print summaries from logistf}
\usage{
logistfPrint(fit, digits = 3)
}
\arguments{
\item{fit}{a fit object returned by \code{logistf}}

\item{digits}{number of digits to round to}
}
\value{
A data frame with a row for each predictor and the
following columns
\item{coef}{log odds ratio}
\item{exp(coef)}{odds ratio}
\item{lower}{lower confidence bound of odds ratio for specified level}
\item{upper}{upper confidence bound of odds ratio for specified level}
\item{p}{p-value}
}
\description{
Print nice summaries from objects returned by \code{logistf}.
}
\details{
There is a lot of raw output from \code{logistf}, and it is not
easy to extract the coefficient table. This function provides
a convenient wrapper to return summaries from \code{logistf}.
}
\examples{
library(logistf)
data(sex2)
fit <- logistf(case ~ age + oc + vic + vicl + vis + dia, data = sex2,
alpha = 0.1)
summary(fit)

## Streamlined summary
logistfPrint(fit)
}
\author{
Samuel Leung, Derek Chiu
}
