% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/maxDate.R
\name{maxDate}
\alias{maxDate}
\title{Maximum of date array}
\usage{
maxDate(t.arr, date.format = MM.DD.YYYY, existing.missing.codes = NA,
  return.missing.code = NA)
}
\description{
Maximum of date array
}

