% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minDateArray.R
\name{minDateArray}
\alias{minDateArray}
\title{Minimum of date array}
\usage{
minDateArray(t.arr, date.format = "MM.DD.YYYY", existing.missing.codes = NA,
  return.missing.code = NA, ...)
}
\arguments{
\item{t.arr}{array of date strings}

\item{date.format}{format of the array of dates}

\item{existing.missing.codes}{missing dates}

\item{return.missing.code}{what to return if there is a missing input}

\item{...}{additional arguments to \code{formatDate}}
}
\description{
Finds the minimum (earliest) date in an array of date strings.
}
\examples{
many.dates <- c("03/21/1992", "04/21/2013", "10/10/2015")
minDateArray(many.dates)

many.dates <- c("2009-03-01", "2010-01-12", "2015-01-11")
minDateArray(many.dates, sep = "-")

ties.dates <- c("2009-03-01", "2009-03-11", "2010-01-12")
minDateArray(ties.dates, sep = "-")
}
\author{
Samuel Leung, Derek Chiu
}

