% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/minWithMissing.R
\name{minWithMissing}
\alias{minWithMissing}
\title{Calculates minimum with specifying missing values}
\usage{
minWithMissing(x, missing.value = -1, return.missing.value = -1)
}
\arguments{
\item{x}{input vector}

\item{missing.value}{missing values in \code{x}}

\item{return.missing.value}{character to return for missing values}
}
\description{
Returns the minimum of input vector, with the ability to specify
which values are missing
}
\note{
\code{NAs} are ignored.
}
\examples{
z <- c(10:1)
minWithMissing(z, c(1, 3))

## All missing
minWithMissing(z, c(1:10))
minWithMissing(z, c(1:10), return.missing.value = "all missing")
}
\author{
Samuel Leung
}

