% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_plot.R
\name{ms_boxplot}
\alias{ms_boxplot}
\title{Boxplot of different MS expression data values}
\usage{
ms_boxplot(x, path, width = 8, height = 10, las = 0)
}
\arguments{
\item{x}{data object returned by \code{ms_process}}

\item{path}{path to save boxplot to. Device is pdf.}

\item{width}{width of plot}

\item{height}{height of plot}

\item{las}{angle of axis labels}
}
\value{
A pdf with three boxplots is saved to disk location specified by
  \code{path}: the raw data values and the log2 and vsn transformed values.
}
\description{
Generate boxplot for raw, log2, and vsn data values
}
\author{
Derek Chiu
}

