% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_plot.R
\name{ms_mean_var}
\alias{ms_mean_var}
\title{Mean-Variance Relationship}
\usage{
ms_mean_var(x, g, title, path, width = 8, height = 10, las = 0)
}
\arguments{
\item{x}{data object returned by \code{ms_process}}

\item{g}{vector of treatment groups}

\item{title}{vector of titles for each treatment group}

\item{path}{path to save boxplot to. Device is pdf.}

\item{width}{width of plot}

\item{height}{height of plot}

\item{las}{angle of axis labels}
}
\value{
A pdf is saved to disk location specified by \code{path}: the vsn
  transformed values and mean-sd plots for each treatment group.
}
\description{
Generate mean-sd plots for the VSN data values.
}
\author{
Derek Chiu
}

