% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_plot.R
\name{ms_plot}
\alias{ms_plot}
\alias{ms_boxplot}
\alias{ms_mean_var}
\title{Plots for MS analyses}
\usage{
ms_boxplot(x, width = 8, height = 10, path = NULL)

ms_mean_var(x, g, title = NULL, width = 8, height = 10, path = NULL)
}
\arguments{
\item{x}{data object returned by \code{ms_process}}

\item{width}{width of plot}

\item{height}{height of plot}

\item{path}{file path to save figure. Device is pdf.}

\item{g}{vector of treatment groups}

\item{title}{vector of titles for each \code{g}}
}
\value{
Both functions return a pdf saved to the file location specified by
\code{path}. \code{ms_boxplot} shows three boxplots of expression values:
raw data values, log2 and vsn transformed values. \code{ms_mean_var} shows
the vsn transformed values and mean-sd plots for each treatment group.
}
\description{
\code{ms_boxplot} shows boxplots of different MS expression data values.
\code{ms_mean_var} shows mean-sd plots for the VSN data values to analyze the
mean-variance relationship.
}
\seealso{
Other Mass Spectrometry functions:
\code{\link[=ms_condition]{ms_condition()}},
\code{\link[=ms_process]{ms_process()}},
\code{\link[=ms_summarize]{ms_summarize()}},
\code{\link[=ms_top]{ms_top()}}
}
\author{
Derek Chiu
}
\concept{Mass Spectrometry functions}
