% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot.R
\name{multiplot}
\alias{multiplot}
\title{Multiple plots}
\usage{
multiplot(..., plotlist = NULL, cols = 1, layout = NULL)
}
\arguments{
\item{...}{pass ggplot objects}

\item{plotlist}{pass a list of ggplot objects}

\item{cols}{number of columns in layout}

\item{layout}{a matrix specifying the layout. If present, \code{cols} is
ignored.}
}
\value{
A grid object made up of multiple ggplots
}
\description{
Place multiple ggplot objects on the same figure space
}
\details{
If the \code{layout} is something like \code{matrix(c(1, 2, 3, 3), nrow = 2, byrow = TRUE)}, then plot 1 will go in the upper left, 2 will go in the upper
right, and 3 will go all the way across the bottom.
}
\author{
Aline Talhouk
}
