% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_cuts.R
\name{name_cuts}
\alias{name_cuts}
\title{Name cutpoint variables}
\usage{
name_cuts(cut.obj)
}
\arguments{
\item{cut.obj}{a factor variable returned by \code{cut2}}
}
\value{
A character string representing the cutpoint variable name.
}
\description{
Create a name for cutpoint variables based on location of cut and number
of groups.
}
\examples{
library(Hmisc)
set.seed(1108)
x <- sample(0:4, size = 1000, replace = TRUE)
c1 <- cut2(x, c(1, 4))
c2 <- cut2(x, c(2, 4))
name_cuts(c1)
name_cuts(c2)
}
\author{
Derek Chiu
}

