% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseCor.R
\name{pairwiseCor}
\alias{pairwiseCor}
\title{Computes all pairwise correlations between the columns of a data frame}
\usage{
pairwiseCor(dataframe)
}
\description{
Args:
 dataframe:A data frame containing numeric variables of interest. 
Returns:
 pairwiseCorDF:  all pairwise absolute correlations,  correlations, Pval, Adj P val by decreasing order of absolute corre,ations.
 @export
}

