% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_rd.R
\name{parse_rd}
\alias{parse_rd}
\title{Parse all Rd files in package}
\usage{
parse_rd(path, tags = c("name", "title", "desc", "details"), save = FALSE)
}
\arguments{
\item{path}{file path to save table}

\item{tags}{sections to extract from Rd files. Missing entries labelled as
\code{NA}}

\item{save}{if \code{TRUE}, the table is saved as a csv}
}
\value{
a table with information on a package's documented objects (functions
  and data). Rows are objects, and columns are tags.
}
\description{
Parse all Rd files in a package and return metadata as a data frame.
}
\details{
The function requires that the working directory is set to the package's root
directory.
}
\examples{
\dontrun{
tab <- parse_rd()
str(tab)
}
}
\author{
Derek Chiu
}

