% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCAplot.R
\name{PCAplot}
\alias{PCAplot}
\title{function to plot the first three PCAs and draw ellipses highlighting differences 
in the levels of a factor}
\usage{
PCAplot(by, tdat, mainT = "", sideT = "")
}
\description{
function to plot the first three PCAs and draw ellipses highlighting differences 
in the levels of a factor
}

