% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotKM.R
\name{plotKM}
\alias{plotKM}
\title{Plot Kaplan-Meier curve}
\usage{
plotKM(input.d, input.formula, main.text, xlab.text, ylab.text, line.name,
  line.color, line.pattern = NULL, line.width = NULL,
  show.test = "single", single.test.type = "logrank",
  round.digits.p.value = 4, obs.survyrs = 10, legend.pos = "bottomleft",
  file.name = "no.file", file.width = 7, file.height = 7,
  grey.scale = FALSE, show.single.test.pos = "default", ...)
}
\description{
Plot Kaplan-Meier curve
}

