% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotKMDetail.R
\name{plotKMDetail}
\alias{plotKMDetail}
\title{Plot detailed Kaplan-Meier curve}
\usage{
plotKMDetail(input.data, surv.formula, main.text, xlab.text, ylab.text,
  line.name, ten.years.surv.95CI, event.count, line.color, obs.survyrs,
  line.pattern = NULL, line.width = NULL, legend.pos = "bottomleft",
  file.name = "no.file", file.width = 7, file.height = 7,
  show.test = "single", single.test.type = "logrank", round.digits.p.value,
  grey.scale = FALSE, show.single.test.pos, ...)
}
\description{
Plot detailed Kaplan-Meier curve
}

