% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSchoenfeld.R
\name{plotSchoenfeld}
\alias{plotSchoenfeld}
\title{Do Schoenfeld residual plot}
\usage{
plotSchoenfeld(input.d, input.formula, vars.to.plot = NULL,
  main = NULL)
}
\arguments{
\item{input.d}{input \code{data.frame}}

\item{input.formula}{survival formula to \code{Surv}}

\item{vars.to.plot}{variable(s) to plot}

\item{main}{title for each variable (i.e. main can be an array)}
}
\description{
Do Schoenfeld residual plot
}
\examples{
# Base output
test1 <- list(time = c(4, 3, 1, 1, 2, 2, 3),
status = c(1, 1, 1, 0, 1, 1, 0),
x = c(0, 2, 1, 1, 1, 0, 0),
sex = c(0, 0, 0, 0, 1, 1, 1))

# Pretty output
plotSchoenfeld(test1, Surv(time, status) ~ x + strata(sex), "x")
}
\author{
Samuel Leung
}
