% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prettyCoxph.R
\name{prettyCoxph}
\alias{prettyCoxph}
\title{Coxph with nicer output}
\usage{
prettyCoxph(input.formula, input.d, use.firth = 1, check.ph = FALSE,
  ph.test.plot.filename = "no.file", ...)
}
\description{
Coxph with nicer output
}

