% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/printCoxMod.R
\name{printCoxMod}
\alias{printCoxMod}
\title{Print Cox model output}
\usage{
printCoxMod(cox, Capt)
}
\arguments{
\item{cox}{an object of class \code{coxph}}

\item{Capt}{a caption}
}
\description{
Prints a Cox model output in a nice HTML table
}
\author{
Aline Talhouk
}

