% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pro}
\alias{pro}
\title{Protein file from Mass Spectrometry experiment}
\format{A data frame with 886 rows and 34 columns}
\usage{
pro
}
\description{
Data for analysis of CLK1 protein between N-Flagged and Untagged Samples
}
\keyword{datasets}
