% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prodWithMissing.R
\name{prodWithMissing}
\alias{prodWithMissing}
\title{Calculates product with specifying missing values}
\usage{
prodWithMissing(x, missing.value = -1, return.missing.value = -1)
}
\arguments{
\item{x}{input vector}

\item{missing.value}{missing values in \code{x}}

\item{return.missing.value}{character to return for missing values}
}
\description{
Returns the product of input vector, with the ability to specify
which values are missing
}
\note{
\code{NAs} are ignored.
}
\examples{
z <- c(10:1)
prodWithMissing(z, c(9, 10))

## All missing
prodWithMissing(z, c(1:10))
prodWithMissing(z, c(1:10), return.missing.value = "all missing")
}
\author{
Samuel Leung
}

