% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{psm}
\alias{psm}
\title{Peptide-Spectrum Match (PSM) file from Mass Spectrometry experiment}
\format{A data frame with 4146 rows and 43 columns}
\usage{
psm
}
\description{
Data for analysis of CLK1 protein between N-Flagged and Untagged Samples
}
\keyword{datasets}
