% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size.R
\name{R2_to_D}
\alias{R2_to_D}
\title{Convert R2 to D}
\usage{
R2_to_D(R2)
}
\arguments{
\item{R2}{R-squared}
}
\value{
The Royston & Sauerbrei D measure
}
\description{
Convert R-Squared to D measure
}
\author{
Derek Chiu
}
