% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratioWithMissing.R
\name{ratioWithMissing}
\alias{ratioWithMissing}
\title{Calculates ratio with specifying missing values}
\usage{
ratioWithMissing(x, y, missing.value = -1, return.missing.value = -1)
}
\arguments{
\item{x}{input vector}

\item{y}{input vector}

\item{missing.value}{missing values in \code{x}}

\item{return.missing.value}{character to return for missing values}
}
\description{
Returns the ratio of input vectors: x/y with the ability to specify which
values are missing
}
\note{
\code{NAs} are ignored.
}
\examples{
x <- c(10:1)
y <- c(1:10)
ratioWithMissing(x, y, c(1:2))

## All missing
ratioWithMissing(x, y, c(1:10))
ratioWithMissing(x, y, c(1:10), return.missing.value = "all missing")
}
\author{
Samuel Leung
}

