% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowColPercent.R
\name{rowColPercent}
\alias{rowColPercent}
\title{Row and column percentages}
\usage{
rowColPercent(t, ...)
}
\arguments{
\item{t}{a matrix}

\item{...}{additional arguments for \code{\link{rowPercent}}
or \code{\link{colPercent}}}
}
\value{
A table with row and column percentages interleaved.
}
\description{
Calculate row and column percentages in a table.
}
\details{
A table with counts, row %, and column % are generated given table t.
The return table will have \code{nrow(t) * 3} rows.
}
\examples{
mat <- matrix(c(5, 3, 8, 9, 2, 4), nrow = 2)
rowColPercent(mat)
rowColPercent(mat, pretty.text = TRUE)
}
\author{
Aline Talhouk, Derek Chiu
}
\seealso{
\code{\link{rowPercent}} for row percentages only,
\code{\link{colPercent}} for column percentages only.
}

