% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowColPercentAsHTML.R
\name{rowColPercentAsHTML}
\alias{rowColPercentAsHTML}
\title{Row and column percentages in HTML format}
\usage{
rowColPercentAsHTML(t, show.count = FALSE, row.names = NULL,
  column.names = NULL, html.table.border = 0, banded.rows = FALSE,
  css.class.name.odd = "odd", css.class.name.even = "even", caption = NA,
  ...)
}
\arguments{
\item{t}{a matrix}

\item{show.count}{logical. If \code{TRUE}, the table counts will be shown.}

\item{row.names}{(optional) vector of row names in table}

\item{column.names}{(optional) vector of column names in table}

\item{html.table.border}{border type for the table. Defaults to 0 in HTML
syntax.}

\item{banded.rows}{logical. If \code{TRUE}, alternating rows will have
different shaded colours.}

\item{css.class.name.odd}{how to format the odd numbered rows in CSS}

\item{css.class.name.even}{how to format the even numbered rows in CSS}

\item{...}{additional arguments to \code{rowColPercent}}
}
\value{
The HTML code used to generate a table that shows row percentages.
}
\description{
Row and column percentages in HTML format
}
\author{
Samuel Leung, Derek Chiu
}

