% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowPercent.R
\name{rowPercent}
\alias{rowPercent}
\title{Row percentages}
\usage{
rowPercent(t, pretty.text = FALSE, keep = TRUE, digits = 4)
}
\arguments{
\item{t}{a matrix}

\item{pretty.text}{logical. If \code{TRUE}, will format the table into nice
display}

\item{keep}{logical. If \code{TRUE}, the original table counts will be kept
along with row percentages.}

\item{digits}{number of digits to round to}
}
\value{
A table with row-wise percentages added. For every row,
the percentages sum to 1.
}
\description{
Calculate row percentages in a table.
}
\details{
Generates a table of row percentages given table \code{t}. Using
\code{pretty.text = TRUE} will add the \% sign to the percentages.
}
\examples{
A <- matrix(c(2, 3, 5, 10), nrow = 2)
rowPercent(A)
rowPercent(A, pretty.text = TRUE)
rowPercent(A, pretty.text = TRUE, keep = TRUE)
}
\author{
Aline Talhouk, Derek Chiu
}

