% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size_prop.R
\name{sample_size_prop}
\alias{sample_size_prop}
\title{Sample Size for Multiple Proportions in One-Way Design}
\usage{
sample_size_prop(props, power = NULL, n = NULL, alpha = 0.05)
}
\arguments{
\item{props}{response proportion in each group}

\item{power}{desired power}

\item{n}{sample size in each group}

\item{alpha}{significance level}
}
\description{
Calculates power and sample size in a one-way design for differences among
response proportions of those groups.
}
\details{
Specify exactly one of \code{power} or \code{n} in order to determine the value of the
other unspecified parameter.
}
\examples{
## Examples from PASS Sample Size Software reference
# Example 1
sample_size_prop(props = c(0.4, 0.2, 0.2), n = 20)
sample_size_prop(props = c(0.4, 0.2, 0.2), n = 40)
sample_size_prop(props = c(0.4, 0.2, 0.2), n = 60)
sample_size_prop(props = c(0.4, 0.2, 0.2), n = 80)
sample_size_prop(props = c(0.4, 0.2, 0.2), n = 100)

# Example 2
sample_size_prop(props = c(0.4, 0.2, 0.2), power = 0.8)
sample_size_prop(props = c(0.4, 0.2, 0.2), power = 0.9)

# Example 3
sample_size_prop(power = 0.9, props = c(0.4, 0.1, 0.1))
sample_size_prop(power = 0.9001, props = c(0.4, 0.2, 0.2))
sample_size_prop(power = 0.9004, props = c(0.4, 0.3, 0.3))
sample_size_prop(power = 0.9038, props = c(0.4, 0.3, 0.1))

# Example 4
sample_size_prop(props = c(0.475, 0.2, 0.2, 0.2), n = 25)
}
\references{
\url{https://www.ncss.com/wp-content/themes/ncss/pdf/Procedures/PASS/Tests_for_Multiple_Proportions_in_a_One-Way_Design.pdf}
}
