% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdFromCI.R
\name{sdFromCI}
\alias{sdFromCI}
\title{Standard deviation of log hazard ratio}
\usage{
sdFromCI(HR, lower.limit, upper.limit, alpha = 0.05)
}
\arguments{
\item{HR}{reported hazard ratio}

\item{lower.limit}{reported lower confidence limit}

\item{upper.limit}{reported upper confidence limit}

\item{alpha}{significance level}
}
\value{
A list with elements
\item{sd}{standard deviation of log HR}
\item{pval}{estimated p-value}
}
\description{
Compute the standard deviation of a log hazard ratio when only given its
repoted confidence limits.
}
\details{
Extract standard deviation from a confidence limit and obtain a corresponding
p-value, which can be compared to the reported p-value. The null p-values are
calculated from the Chi-squared 1 distribution.
}
\examples{
library(meta)
studlab <- c("PORTEC", "Leuven", "TCGA", "Billingsley")
effects <- log(c(0.43, 0.18, 0.12, 0.37))
se_effects <- c(sdFromCI(0.43, 0.13, 1.37)$sd,
                sdFromCI(0.18, 0.01, 3.11)$sd,
                sdFromCI(0.12, 0.01, 2.11)$sd,
                sdFromCI(0.37, 0.09, 1.54)$sd)
metagen(effects, se_effects, studlab = studlab, sm = "HR", comb.fixed = TRUE)
}
\references{
http://www.ncbi.nlm.nih.gov/pmc/articles/PMC1920534/
}
\author{
Aline Talhouk, Derek Chiu
}
