% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sem.R
\name{sem}
\alias{sem}
\title{Standard error of the mean}
\usage{
sem(x, missing.value = NA, return.missing.value = NA)
}
\description{
Standard error of the mean
}
\author{
Samuel Leung
}
\references{
http://en.wikipedia.org/wiki/Standard_error
}

