% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupKnitr.R
\name{setupKnitr}
\alias{setupKnitr}
\title{Setup knitr}
\usage{
setupKnitr(fw = 8, fh = 6)
}
\arguments{
\item{fw}{figure width (default = 8)}

\item{fh}{figure height (default = 6)}
}
\description{
Setup knitr
}
