% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_string.R
\name{simpleCap}
\alias{simpleCap}
\title{Simple capitalization}
\usage{
simpleCap(x, first.only = FALSE)
}
\arguments{
\item{x}{character string}

\item{first.only}{logical; if \code{TRUE}, only the first word will be
capitalized}
}
\value{
A character string with every word's first letter capitalized.
}
\description{
Capitalize the first letter of every word in a character string.
}
\details{
To capitalize only the first word, use \code{first.only = TRUE}.
}
\examples{
simpleCap("clear cell")
simpleCap("high grade serous carcinoma")
simpleCap("ovarian cancer", first.only = TRUE)
}
\references{
http://stackoverflow.com/questions/6364783/capitalize-the-first-letter-of-both-words-in-a-two-word-string
}
\author{
Samuel Leung
}
