% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size.R
\name{ssize_D}
\alias{ssize_D}
\title{Sample Size Calculation}
\usage{
ssize_D(
  fit,
  D = NULL,
  cens = NULL,
  p = 0.1,
  alpha = 0.05,
  power = 0.8,
  delta = 0.25,
  trial = c("superiority", "non-inferiority")
)
}
\arguments{
\item{fit}{an object fit of class \code{coxph}}

\item{D}{discrimination measure D, defaults to output of
\code{\link[survival:royston]{survival::royston()}} for \code{fit}}

\item{cens}{censoring proportion, defaults to proportion calculated from
\code{fit}}

\item{p}{proportion of \code{D} accepted as \code{delta}, defaults to 10\%.}

\item{alpha}{Type I error}

\item{power}{statistical power (1 - Type II error)}

\item{delta}{superiority/non-inferiority margin. The difference we wish to
detect between a prior model and a new model.}

\item{trial}{either "superiority" or "non-inferiority"}
}
\value{
The final sample size needed in the new model using the supplied
input parameters.
}
\description{
Sample size calculation using model-based discrimination measure.
}
\details{
The sample size calculation uses a discrimination measured based on
multivariable prognostic time-to-event models.
}
\references{
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4603804/}
}
\author{
Derek Chiu
}
