% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsWithMissing.R
\name{statsWithMissing}
\alias{statsWithMissing}
\alias{minWithMissing}
\alias{maxWithMissing}
\alias{meanWithMissing}
\alias{sumWithMissing}
\alias{prodWithMissing}
\alias{ratioWithMissing}
\title{Calculates statistics with specifying missing values}
\usage{
minWithMissing(x, missing.value = -1, return.missing.value = -1)

maxWithMissing(x, missing.value = -1, return.missing.value = -1)

meanWithMissing(x, missing.value = -1, return.missing.value = -1)

sumWithMissing(x, missing.value = -1, return.missing.value = -1)

prodWithMissing(x, missing.value = -1, return.missing.value = -1)

ratioWithMissing(x, y, missing.value = -1, return.missing.value = -1)
}
\arguments{
\item{x, y}{input vector}

\item{missing.value}{missing values in \code{x}}

\item{return.missing.value}{character to return for missing values}
}
\description{
Returns various statistics, with the ability to specify which values are
missing.
}
\details{
The statistics supported include \code{min, max, mean, sum, prod}, as
indicated by each function's prefix. In addition, \code{ratioWithMissing}
calculates the ratio \code{x / y}.
}
\note{
\code{NAs} are ignored.
}
\examples{
x <- c(10:1)
y <- c(1:10)
z <- c(10:1)
minWithMissing(z, c(1, 3))
maxWithMissing(z, c(9, 10))
meanWithMissing(z, c(3:7))
sumWithMissing(z, c(9, 10))
prodWithMissing(z, c(9, 10))
ratioWithMissing(x, y, c(1:2))

## All missing
minWithMissing(z, c(1:10))
minWithMissing(z, c(1:10), return.missing.value = "all missing")
maxWithMissing(z, c(1:10))
maxWithMissing(z, c(1:10), return.missing.value = "all missing")
meanWithMissing(z, c(1:10))
meanWithMissing(z, c(1:10), return.missing.value = "all missing")
prodWithMissing(z, c(1:10))
prodWithMissing(z, c(1:10), return.missing.value = "all missing")
}
\author{
Samuel Leung, Derek Chiu
}
