% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summaryAsHTML.R
\name{summaryAsHTML}
\alias{summaryAsHTML}
\title{Summary table in HTML format}
\usage{
summaryAsHTML(d)
}
\arguments{
\item{d}{assume \code{d} is an array of numbers}
}
\value{
summary table with annotated HTMl code
}
\description{
Generate summary table as an HTML table
}
\author{
Samuel Leung
}

