% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStatsBy.R
\name{SummaryStatsBy}
\alias{SummaryStatsBy}
\title{Generate cohort characteristics}
\usage{
SummaryStatsBy(data, by1, by2, var.names, stats = c("mean", "sd", "median",
  "IQR", "range", "missing"), marker.description = by1,
  groups.description = by2, var.descriptions = var.names, digits = 3,
  missing.codes.highlight = NULL, missing.codes = c("N/A", "", "Unk"),
  decimal = 0, caption = NA)
}
\arguments{
\item{data}{The \code{data.frame} containing the data}

\item{by1}{split variable 1}

\item{by2}{split variable 2}

\item{var.names}{The variables that you want the statistics for}

\item{stats}{statistics to show compute for continuous variables}

\item{marker.description}{The description for the variable(s) to split}

\item{groups.description}{The groupings of the marker}

\item{var.descriptions}{descriptions of \code{var.names}}

\item{digits}{number of digits to round to}

\item{missing.codes.highlight}{NAs}

\item{missing.codes}{search strings to detect missing entries}

\item{decimal}{number of decimal places}

\item{caption}{caption for summary table}
}
\description{
Generate cohort characteristics
}
\author{
Aline Talhouk, Derek Chiu
}

