% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumWithMissing.R
\name{sumWithMissing}
\alias{sumWithMissing}
\title{Calculates maximum with specifying missing values}
\usage{
sumWithMissing(x, missing.value = -1, return.missing.value = -1)
}
\arguments{
\item{x}{input vector}

\item{missing.value}{missing values in \code{x}}

\item{return.missing.value}{character to return for missing values}
}
\description{
Returns the sum of input vector, with the ability to specify
which values are missing
}
\note{
\code{NAs} are ignored.
}
\examples{
z <- c(10:1)
sumWithMissing(z, c(9, 10))

## All missing
sumWithMissing(z, c(1:10))
sumWithMissing(z, c(1:10), return.missing.value = "all missing")
}
\author{
Samuel Leung
}

