% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Xunivcoxph.R
\name{Xunivcoxph}
\alias{Xunivcoxph}
\title{Univariate cox proprtional hazards model}
\usage{
Xunivcoxph(mod, coxph.type = "coxph", digits = 3)
}
\value{
Hazard ratio, confidence interval, and likelihood ratio test
p-value
}
\description{
Univariate cox proprtional hazards model
}
\author{
Aline Talhouk
}

