% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTableNumber.R
\name{addTableNumber}
\alias{addTableNumber}
\title{Add table number}
\usage{
addTableNumber(caption, table.counter.str.default = "Table \%s: ")
}
\arguments{
\item{caption}{table caption}

\item{table.counter.str.default}{the default string preceding the table count}
}
\value{
A caption with a table counter prepended to its text.
}
\description{
Prepends a table counter label to a table caption.
}
\details{
The table counter automatically increments based on the number of existing
tables that have already called  using \code{addTableNumber}.
}
\examples{
addTableNumber("Overall Survival Hazard Ratios")
addTableNumber("Correlation Heatmap", table.counter.str.default = "Figure \%s:
")
}
\author{
Samuel Leung
}
