% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addToDate.R
\name{addToDate}
\alias{addToDate}
\title{Add to a date}
\usage{
addToDate(org.date, delta, date.format = "MM.DD.YYYY", units = "days",
  existing.missing.codes = NA, return.missing.code = NA, ...)
}
\arguments{
\item{org.date}{original date}

\item{delta}{amount of time to add to \code{org.date}}

\item{date.format}{how to format the resulting date}

\item{units}{the unit of time for which to take the difference. Defaults to
"days".}

\item{existing.missing.codes}{missing dates}

\item{return.missing.code}{what to return if there is a missing input}

\item{...}{additional arguments to \code{formatDate}}
}
\value{
The new date after adding \code{delta} to \code{org.date}.
}
\description{
Adds a period of time to a date string and returns a new date string.
}
\details{
The added period of time can be in \code{"days"}, \code{"months"}, or
\code{"years"}. If \code{delta} is negative, then the returned date will be
earlier than \code{org.date}. The output date format will be the same as the
input date format.
}
\examples{
addToDate("2014/07/08", 10, date.format = "YYYY.MM.DD")
addToDate("2014-07-08", 10, date.format = "YYYY.MM.DD", sep = "-")
addToDate("2014/07/08", 10, date.format = "YYYY.MM.DD", units = "months")
addToDate("2014/07/08", -10, date.format = "YYYY.MM.DD", units = "years")
}
\author{
Samuel Leung
}

