% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/addToDate.R
\name{addtoDate}
\alias{addtoDate}
\title{Add to a date}
\usage{
addtoDate(org.date, delta, date.format = MM.DD.YYYY, units = "days",
  existing.missing.codes = NA, return.missing.code = NA)
}
\description{
Add to a date
}

