% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessSurvTime.R
\name{assessSurvTime}
\alias{assessSurvTime}
\title{Assess survival time}
\usage{
assessSurvTime(T1, T2, status)
}
\arguments{
\item{T1}{vector of start dates}

\item{T2}{vector of end dates}

\item{status}{logical;}
}
\value{
A list with elements
\item{Otime}{Observation time}
\item{Stime}{Censoring time}
\item{Etime}{Time to end of study}
\item{KFT}{Known Function Time}
\item{RevKM}{Reverse Kaplan-Meier Time}
}
\description{
Given range of survival times and censoring status, provides different 
time-related summary statistics
}
\details{
The observation time is defined as the median time in days between all 
\code{T1} and \code{T2}. Censoring time is the median time in days between 
all \code{T1} and \code{T2} for events only.
}
\author{
Samuel Leung
}

