% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_annotated.R
\name{barplotSum}
\alias{barplotSum}
\title{Annotated Barplot}
\usage{
barplotSum(tx, ttl = "")
}
\arguments{
\item{tx}{table of a categorical variable}

\item{ttl}{the title of the boxplot}
}
\description{
Function to provide a barplot from a table of categorical variable with
labels wrapped
}
\author{
Aline Talhouk
}
