% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binaryCM.R
\name{binaryCM}
\alias{binaryCM}
\title{Confusion matrix summaries}
\usage{
binaryCM(
  x,
  y,
  seed = 20,
  num.boot = 1000,
  pcond = 1,
  conf.level = 0.95,
  digits = 4,
  method = "wilson",
  verbose = FALSE
)
}
\arguments{
\item{x}{a vector of reference classes}

\item{y}{a vector of predicted classes}

\item{seed}{random seed for bootstrapping}

\item{num.boot}{number of times to bootstrap. Defaults to 1000.}

\item{pcond}{a string or value to be considered positive condition; defaults
to 1.}

\item{conf.level}{confidence level. Defaults to 95\%.}

\item{digits}{number of digits to round summaries to}

\item{method}{method for obtaining confidence intervals for binomial
probabilities. See \code{\link[Hmisc:binconf]{Hmisc::binconf()}} for details.}

\item{verbose}{logical; if \code{TRUE}, outputs are printed to the screen}
}
\value{
A confusion matrix for the predicted and reference classes. Then
the estimated statistics along with bootstrapped confidence intervals. A
list with the following elements
\item{CM}{The confusion matrix, whose columns are the predicted conditions
and its rows are the true conditions}
\item{Accuracy}{Accuracy point estimate, lower bound and upper bound for
bootstrapped CI}
\item{Sensitivity}{Sensitivity point estimate, lower bound and upper bound
for bootstrapped CI}
\item{Specificity}{Specificity point estimate, lower bound and upper bound
for bootstrapped CI}
\item{PPV}{PPV point estimate, lower bound and upper bound for
bootstrapped CI}
\item{NPV}{NPV point estimate, lower bound and upper bound for
bootstrapped CI}
\item{kappa}{kappa point estimate, lower bound and upper bound for
bootstrapped CI}
\item{table}{a data frame that contains all 6 of the estimated statistics
along with confidence intervals}
}
\description{
Calculates summaries from cross-tabulated reference and prediction labels for
a two-class variable.
}
\details{
Given two dichotomous variables summarized in a confusion matrix, this
function provides performance summaries. The accuracy, sensitivity,
specificity, positive predictive value (PPV), negative predictive value
(NPV), and the kappa statistic, along with their bootstrapped confidence
intervals are returned.

Note that the classes given in \code{x} and \code{y} must be binary.
}
\examples{
### 95\% CI from 1000 bootstraped samples
set.seed(547)
n <- 80
x <- rbinom(n, size = 1, prob = 0.6)
y <- rbinom(n, size = 1, prob = 0.4)
binaryCM(x, y)

### 90\% CI from 500 bootstrapped samples
binaryCM(x, y, num.boot = 500, conf.level = 0.90)

### Round to 2 digits
binaryCM(x, y, digits = 2)
}
\seealso{
Other confusion matrix functions:
\code{\link[=binaryCMAsHTML]{binaryCMAsHTML()}},
\code{\link[=multiClassCM]{multiClassCM()}}
}
\author{
Aline Talhouk, Derek Chiu
}
\concept{confusion matrix functions}
