% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binaryCMAsHTML.R
\name{binaryCMAsHTML}
\alias{binaryCMAsHTML}
\title{Confusion matrix results in HTML}
\usage{
binaryCMAsHTML(x, y, ref.description = NULL, digits = 4, seed = 20,
  num.boot = 1000, conf.level = 0.95, show.ci = TRUE)
}
\arguments{
\item{x}{vector of reference classes}

\item{y}{vector of predicted classes}

\item{ref.description}{description of classes}

\item{digits}{number of digits to round p-values to}

\item{seed}{random seed for bootstrap resampling}

\item{num.boot}{number of bootstrap confidence intervals}

\item{conf.level}{confidence level. Defaults to 95\%.}

\item{show.ci}{if \code{TRUE} (default), the confidence intervals are shown.}
}
\value{
A character string that can be parsed as HTML code to display a nice
confusion matrix summary.
}
\description{
Prints results from \code{\link[=binaryCM]{binaryCM()}} into a nice HTML table format.
}
\examples{
# 95\% CI from 5 bootstraped samples
library(htmlTable)
set.seed(547)
n <- 20
x <- rbinom(n, size = 1, prob = 0.6)
y <- rbinom(n, size = 1, prob = 0.4)
results <- binaryCMAsHTML(x, y, "Test", num.boot = 1000)
htmlTable(results)

results.no.ci <- binaryCMAsHTML(x, y, "Test", num.boot = 1000, show.ci =
FALSE)
htmlTable(results.no.ci)
}
\seealso{
Other confusion matrix functions: \code{\link{binaryCM}},
  \code{\link{multiClassCM}}
}
\author{
Samuel Leung, Derek Chiu
}
\concept{confusion matrix functions}
