% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biostatUtil.R
\docType{package}
\name{biostatUtil}
\alias{biostatUtil-package}
\alias{biostatUtil}
\title{biostatUtil: utility functions for biostatistical projects}
\description{
biostatUtil provides a wide variety of data analysis, visualization, and
presentation functions used in biostatistical projects.
}
\details{
This package is currently in development stage while still lacking
comprehensive documentation and testing. The goal is to ultimately separate
biostatUtil into smaller, more cohesive packages each with better defined
purposes.

Functions can be categorized into one of the following: \itemize{
\item Survival analysis
\item Character manipulation
\item Date manipulation
\item Plotting survival-related figures
\item Miscellaneous functions
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://talhouklab.github.io/biostatUtil/}
}

}
\author{
\strong{Maintainer}: Derek Chiu \email{dchiu@bccrc.ca}

Authors:
\itemize{
  \item Aline Talhouk \email{atalhouk@bccrc.ca}
  \item Samuel Leung \email{Samuel.Leung@vch.ca}
}

}
