% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootMean.R
\name{bootMean}
\alias{bootMean}
\title{Mean and bootstrap confidence interval}
\usage{
bootMean(x, num.boot = 1000, conf.level = 0.95, seed = 12, ...)
}
\arguments{
\item{x}{a vector (or matrix)}

\item{num.boot}{number of bootstrap samples. Defaults to 1000.}

\item{conf.level}{confidence level.}

\item{seed}{random seed for resampling}

\item{...}{additional arguments to \code{mean}}
}
\value{
A list with elements
\item{obs.mean}{mean}
\item{ci}{bootstraped confidence interval}
\item{n}{number of bootstrap samples used}
}
\description{
Finds mean of a vector, with bootstrapped confidence bounds.
}
\details{
Takes a numeric vector and resamples with replacement \code{num.boot} times.
If the input vector has any \code{NA} entries, include the argument
\code{na.rm = TRUE} from \code{mean}.
}
\examples{
## Vectors
set.seed(344)
bootMean(rnorm(100, 5, 3))
bootMean(rnorm(100, 5, 3), num.boot = 500)
bootMean(rnorm(100, 4, 3), conf.level = 0.90)

## Missing Values
set.seed(344)
s <- ifelse(rexp(100, 0.5) < 1, NA, rexp(100, 0.5))
bootMean(s)  # doesn't work
bootMean(s, na.rm = TRUE)
}
\author{
Aline Talhouk, Derek Chiu
}
