% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_cuts.R
\name{build_cuts}
\alias{build_cuts}
\title{Build cutpoint variables}
\usage{
build_cuts(x, n = c("b", "t", "qd", "qn"), var.prefix = NULL,
  list = FALSE)
}
\arguments{
\item{x}{vector to split by certain cutpoints}

\item{n}{either "b" for binarization (2 groups), "t" for trinarization (3
groups), "q" for quads (4 groups), or "qn" for quints (5 groups)}

\item{var.prefix}{variable name prefix}

\item{list}{if \code{TRUE}, the variables are returned as a list.}
}
\value{
By default, a tibble of cutpoint variables built from a categorical
biomarker. The number of columns correspond to all the ways the biomarker
could be cut into \code{n} bins. Each column name starts with a "b", "t",
"qd", or "qn" for "binarization", "trinarization", "quads", or "quints",
respectively, with the levels being compared separated by "v". If
\code{list = FALSE}, each cutpoint variable is an element of a list.
}
\description{
Transforms an ordinal variable into anywhere from two to five groups for
cutpoint analysis.
}
\examples{
set.seed(1108)
x <- sample(0:4, size = 1000, replace = TRUE)
build_cuts(x, n = "b")
build_cuts(x, n = "t")
build_cuts(x, n = "t", var.prefix = "PHGDH")
str(build_cuts(x, n = "qd", list = TRUE))
}
\author{
Derek Chiu
}
