% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_format.R
\name{cleanDate}
\alias{cleanDate}
\title{Clean dates}
\usage{
cleanDate(x, original.format, preferred.format,
  existing.missing.codes = "", return.missing.code = NA, sep = "/")
}
\arguments{
\item{x}{a date string or a numeric representation of a date
(e.g. January 13th, 1991 would be 19910113)}

\item{original.format}{format of input \code{x}}

\item{preferred.format}{format to change \code{x} to}

\item{existing.missing.codes}{missing dates}

\item{return.missing.code}{what to return if there is a missing input}

\item{sep}{date separator. Defaults to "/"}
}
\value{
A date string cleaned and formatted from the original (unformatted)
date
}
\description{
Clean a date and reformat it to another style.
}
\examples{
cleanDate("09/11/1991", original.format = "MM.DD.YYYY", preferred.format = "DD.MM.YYYY")
cleanDate(11091991, original.format = "DD.MM.YYYY", preferred.format = "YYYY.MMM.DD")
cleanDate(11091991, original.format = "DD.MM.YYYY", preferred.format = "YYYY.MMM.DD", sep = "-")
}
\seealso{
Other date formatting functions: \code{\link{formatDate}},
  \code{\link{getFormat}}, \code{\link{numericToDate}}
}
\author{
Samuel Leung, Derek Chiu
}
\concept{date formatting functions}
