% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/colPercent.R
\name{colPercent}
\alias{colPercent}
\title{Column-wise percentages}
\usage{
colPercent(t, pretty.text = FALSE, keep = FALSE, digits = 4)
}
\arguments{
\item{t}{a matrix}

\item{pretty.text}{logical indicating whether to format the table}

\item{keep}{logical indicating whether to keep the raw table values}

\item{digits}{number of digits to round to}
}
\value{
A table with column percentages added.
}
\description{
Generates a table of column percentages given a table.
}
\details{
Details to be filled.
}
\author{
Aline Talhouk, Derek Chiu
}

