% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compareDate.R
\name{compareDate}
\alias{compareDate}
\title{Compare two dates}
\usage{
compareDate(d1, d2, date.format = "MM.DD.YYYY", units = "days",
  existing.missing.codes = NA, return.missing.code = NA, ...)
}
\arguments{
\item{d1}{first date}

\item{d2}{second date}

\item{date.format}{how to format the resulting date}

\item{units}{the unit of time for which to take the difference. Defaults to "days".}

\item{existing.missing.codes}{missing dates}

\item{return.missing.code}{what to return if there is a missing input}

\item{...}{additional arguments to \code{formatDate}}
}
\value{
Returns 1 if \code{d1 > d2}, -1 if \code{d1 < d2}, and 0 if \code{d1 == d2}.
}
\description{
Compares the difference between two dates.
}
\details{
Calls \code{diffDate} and returns integer values specifying which date is earlier
(or if they are the same).
}
\examples{
compareDate("01/22/1949", "04/13/1950", date.format = "MM.DD.YYYY")
compareDate("04/13/1950", "04/13/1950", date.format = "MM.DD.YYYY")
compareDate("04/13/1959", "04/13/1950", date.format = "MM.DD.YYYY")
}
\author{
Samuel Leung, Derek Chiu
}

