% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/defineEventDate.R
\name{defineEventDate}
\alias{defineEventDate}
\title{Define event date}
\usage{
defineEventDate(ed, cutoff, event = "OS", format.dates.in.ed = TRUE,
  cutoff.date.format = "\%Y-\%m-\%d", status.only = FALSE)
}
\description{
Define event date
}
\author{
Samuel Leung, Aline Talhouk
}

