% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffDate.R
\name{diffDate}
\alias{diffDate}
\title{Difference between two dates}
\usage{
diffDate(d1, d2, date.format = "MM.DD.YYYY", units = c("days", "weeks",
  "months", "years"), existing.missing.codes = NA, return.missing.code = NA,
  ...)
}
\arguments{
\item{d1}{first date}

\item{d2}{second date}

\item{date.format}{how to format the resulting date}

\item{units}{the unit of time for which to take the difference. Defaults to 
"days".}

\item{existing.missing.codes}{missing dates}

\item{return.missing.code}{what to return if there is a missing input}

\item{...}{additional arguments to \code{formatDate}}
}
\value{
The difference between two dates \code{d1 - d2} returned in the 
  specified unit of time.
}
\description{
Takes the difference between two dates and returns in the specified unit of 
time.
}
\details{
The unit of time can be \code{"days"} (default), \code{"weeks"},
\code{"months"}, or \code{"years"}.
}
\examples{
## Wrong Order
diffDate("1992/01/27", "2003/03/21", date.format = "YYYY.MM.DD")

## Later date comes first, subtracts the second date
diffDate("2003/03/21", "1992/01/27", date.format = "YYYY.MM.DD")

## Different separator
diffDate("2003-03-21", "1992-01-27", date.format = "YYYY.MM.DD", sep = "-")
}
\author{
Samuel Leung, Derek Chiu
}

