% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_do.R
\name{doBarplot}
\alias{doBarplot}
\title{Barplot with counts}
\usage{
doBarplot(data, var, xlab = var, ylab = "Frequency", title = NULL,
  digits = 3)
}
\arguments{
\item{data}{data.frame with column names}

\item{var}{string of variable in \code{data} to graph on}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{title}{barplot title}

\item{digits}{number of digits to round to}
}
\value{
Barplot of frequencies for each group
}
\description{
Generates barplot on variables and shows counts per group.
}
\details{
The number of data points in each group is reported.
The number of scorable and missing data are also reported by number
and percentage in the barplot title.
}
\note{
Function expects missing to be \code{NA}. Do not filter out missing data
as this function reports missing data frequencies.
}
\examples{
doBarplot(mtcars, "cyl", "title = Number of cylinders")
doBarplot(mtcars, "cyl", "Cylinders", title = "Number of cylinders")
}
\author{
Samuel Leung, Derek Chiu
}
