% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doCheckPH.R
\name{doCheckPH}
\alias{doCheckPH}
\alias{doCheckPHMulti}
\title{Check Cox proportional hazards assumption}
\usage{
doCheckPH(data, var.name, var.description)

doCheckPHMulti(data, var.name, var.description)
}
\arguments{
\item{data}{the data object}

\item{var.name}{column name of variable to test in \code{data}}

\item{var.description}{label to use for \code{var.name} in plots}
}
\value{
Schoenfeld residual plots for each survival outcome
}
\description{
Use Schoenfeld residuals to test the Cox proportional hazards assumption in a
(multivariable) survival model.
}
\details{
\code{doCheckPH} is intended for use in univariable models, and
\code{doCheckPHMulti} is intended for use in multivariable models.
}
\author{
Samuel Leung
}
