% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doCheckPHMulti.R
\name{doCheckPHMulti}
\alias{doCheckPHMulti}
\title{Check the proportional hazards assumption in multivariable model}
\usage{
doCheckPHMulti(data, var.name, var.description)
}
\arguments{
\item{data}{the data object}

\item{var.name}{column name of variable to test in \code{data}}

\item{var.description}{label to use for \code{var.name} in plots}
}
\value{
Schoenfeld residual plots for each survival outcome
}
\description{
Use Schoenfeld residuals to test the Cox proportional hazards assumption.
}
\author{
Samuel Leung
}

