% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doCohortCharacteristics.R
\name{doCohortCharacteristics}
\alias{doCohortCharacteristics}
\title{Generate cohort characteristics}
\usage{
doCohortCharacteristics(input.d, marker.name, marker.description,
  var.names, is.var.continuous, var.descriptions,
  marker.value.labels.tolower = TRUE, show.missing = TRUE,
  show.missing.continuous = TRUE, do.droplevels = TRUE,
  show.percent = "both", stat.tests = NULL,
  chisq.test.simulate.p.value = FALSE,
  stat.test.column.header = "association/correlation test",
  round.digits.p.value = 4, num.boot = 1000,
  missing.codes.highlight = NULL, missing.codes = c("N/A", "", "Unk"),
  decimal = 0, caption = NA, html.table.border = 0,
  banded.rows = FALSE, css.class.name.odd = "odd",
  css.class.name.even = "even", custom.marker.labels = NULL,
  custom.total.label = NULL, split.table = 200, ...)
}
\arguments{
\item{input.d}{The \code{data.frame} containing the data}

\item{marker.name}{The variable that you want to split into different columns}

\item{marker.description}{The description for the variable(s) to split}

\item{var.names}{The variables that you want the statistics for}

\item{is.var.continuous}{Vector of length equal to the length of var.names
with 1 indicating a continuous variable and 0 otherwise (this should be
inferred in the function)}

\item{var.descriptions}{Vector of strings to describe the variables as they
are to appear in the table}

\item{marker.value.labels.tolower}{Indicator as to whether to put marker
value labels to lower case}

\item{show.missing}{an indicator to whether to show missing values}

\item{show.missing.continuous}{if set to \code{FALSE} and \code{show.missing == FALSE}, will not show the number of missing cases for continuous
variables. Otherwise, it shows the number of missing for continuous
variables even if \code{show.missing == FALSE}.}

\item{do.droplevels}{drop categories of unobserved levels set to \code{TRUE}}

\item{show.percent}{defaults to "both" which shows both rows and columns
other possible values: "column", "row".}

\item{stat.tests}{statistical test to perform. \code{NULL} indicates do not
do test for all variables, \code{NA} indicates do not do test for specified
variable. Tests: chisq, fisher, ttest, wilcox, kendall, spearman, pearson,
kruskal, confusionMarkerAsRef, confusionVarAsRef}

\item{chisq.test.simulate.p.value}{Whether to simulate p-value for chi-square
test. this parameter is ignored if chi-square is not used.  Default
value=FALSE}

\item{stat.test.column.header}{The name to show on the header defaults to
"association/correlation test"}

\item{round.digits.p.value}{The number of digits to round the P values}

\item{num.boot}{the number of bootstrap samples for any bootstrap method that
may be used}

\item{missing.codes.highlight}{default to \code{NULL} this indicates whether
we wanted the missing values broken down down or lumped together.}

\item{missing.codes}{a vector to indicate how missing values are coded,
default is \code{c("N/A", "", "Unk")}}

\item{decimal}{number of decimal places to show for aggregate numbers such as
proportions or averages; default to 0}

\item{caption}{caption to use for the Table}

\item{html.table.border}{the border type to use for html tables}

\item{banded.rows}{If \code{TRUE}, rows have alternating shading colour}

\item{css.class.name.odd}{Used to set the row colour for odd rows}

\item{css.class.name.even}{Used to set the row colour for even rows}

\item{custom.marker.labels}{labels of marker to show; default \code{NULL}
means using existing value label of the marker}

\item{custom.total.label}{label of the "Total" column; default \code{NULL}
means show "Total"}

\item{split.table}{number of chars per row before table is split.}

\item{...}{additional arguments to \code{pander}}
}
\value{
A table with statistics reported for multiple variables, such as
mean, median, and range for continuous variables and proportions and
percentages for categorical variables. Relevant association and correlation
tests are performed as well.
}
\description{
Generate cohort characteristics
}
\examples{
dcc <- doCohortCharacteristics( input.d = mtcars, marker.name = "cyl",
marker.description = "cylinders", var.names = c("disp", "hp"),
var.descriptions = c("displacement", "horsepower"), is.var.continuous =
c(TRUE, TRUE), caption = "Some mtcars summaries")
htmlTable::htmlTable(dcc$result.table.html)
}
\author{
Aline Talhouk
}
