% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/doCohortCharacteristics.R
\name{doCohortCharacteristics}
\alias{doCohortCharacteristics}
\title{Generate cohort characteristics}
\usage{
doCohortCharacteristics(input.d, marker.name, marker.description, var.names,
  is.var.continuous, var.descriptions, marker.value.labels.tolower = TRUE,
  show.missing = TRUE, show.missing.continuous = TRUE,
  do.droplevels = TRUE, show.percent = "both", stat.tests = NULL,
  stat.test.column.header = "association/correlation test",
  round.digits.p.value = 4, num.boot.for.ci = 1000,
  missing.codes.highlight = NULL, missing.codes = c("N/A", "", "Unk"),
  decimal = 0, caption = NA, html.table.border = 0, banded.rows = FALSE,
  css.class.name.odd = "odd", css.class.name.even = "even")
}
\description{
Generate cohort characteristics
}

