% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_do.R
\name{doHist}
\alias{doHist}
\title{Do histogram with median}
\usage{
doHist(data, var, xlab = var, title = NULL, show.title = TRUE,
  br = ceiling(sqrt(nrow(data))), digits = 3, score.lab = "scorable",
  ...)
}
\arguments{
\item{data}{data.frame with column names}

\item{var}{string of variable in \code{data} to graph on}

\item{xlab}{x-axis label}

\item{title}{histogram title}

\item{show.title}{logical. If \code{TRUE} (default), the title is shown}

\item{br}{breaks in histogram. By default, the number of bins is taking as
the ceiling of the square root of the number of rows in \code{data}}

\item{digits}{number of digits to round for median, Q1, Q3}

\item{score.lab}{label for non-missing cases}

\item{...}{additional arguments to \code{hist}}
}
\value{
A histogram with some annotated values.
}
\description{
Plot a histogram with the median, first quartile, and third quartile reported.
}
\details{
Expects missing to be \code{NA}. Do not filter out missing data
as this function reports missing data counts.
}
\examples{
doHist(mtcars, "mpg")
doHist(mtcars, "mpg", "MPG")
doHist(mtcars, "mpg", title = "Distribution of MPG")
}
\author{
Samuel Leung, Derek Chiu
}
