% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/doHist.R
\name{doHist}
\alias{doHist}
\title{Do histogram with median}
\usage{
doHist(input.d, data.description, biomarker.var.name, biomarker.name,
  show.title = TRUE, br = 100, digits = 3,
  interpretable.label = "scorable")
}
\description{
Do histogram with median
}

