% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doKMPlots.R
\name{doKMPlots}
\alias{doKMPlots}
\title{Make Kaplan-Meier plots}
\usage{
doKMPlots(input.d, time, status, var.name, var.description,
  surv.type = c("os", "dss", "pfs"), shading.colors = c("blue2",
  "red2", "deepskyblue", "indianred3"), line.name = NULL,
  line.pattern = NULL, legend = FALSE, cox.ref.group = NULL,
  use.firth = -1, CI = TRUE, HR = TRUE, show.risk = TRUE,
  km.plot.ref.group = "single", single.test.type = "logrank",
  use.ggkm = FALSE, ...)
}
\arguments{
\item{input.d}{\code{data.frame} containing data}

\item{time}{follow up time}

\item{status}{status indicator}

\item{var.name}{name of variable to make Kaplan-Meier plots on}

\item{var.description}{description for \code{var.name}}

\item{surv.type}{survival outcome. Either "os", "dss", or "pfs".}

\item{shading.colors}{colors for survival curves}

\item{line.name}{names for each survival curve}

\item{line.pattern}{line type for survival curves}

\item{legend}{logical; if \code{TRUE}, the legend is overlaid on the graph
(instead of on the side).}

\item{cox.ref.group}{specify reference group for cox model i.e. hazard
ratio(s)}

\item{use.firth}{Whether to use Firth's correction for plotting the curves}

\item{CI}{logical; if \code{TRUE}, will plot confidence bands}

\item{HR}{logical; if \code{TRUE}, will show hazard ratios}

\item{show.risk}{logical; if \code{TRUE}, will show the number of people at
risk at each time of death beneath the plot}

\item{km.plot.ref.group}{specify KM plot reference group; "single" means a
lump log-rank statistic}

\item{single.test.type}{test to use for survival curves. Defaults to
"logrank".}

\item{use.ggkm}{if \code{TRUE}, will use function \code{ggkm} for plotting}

\item{...}{additional arguments to other functions and methods}
}
\value{
A Kaplan-Meier plot for the specified survival outcome split on the
desired variable.
}
\description{
Make Kaplan-Meier plots
}
\author{
Samuel Leung, Derek Chiu
}
