% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/doKMPlotsPOLE.R
\name{doKMPlotsPOLE}
\alias{doKMPlotsPOLE}
\title{Make Kaplan-Meier plots for POLE}
\usage{
doKMPlotsPOLE(data, surv.type, use.aline.plot = FALSE, ...)
}
\description{
Make Kaplan-Meier plots for POLE
}

