% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/formatDate.R
\name{formatDate}
\alias{formatDate}
\title{Format a date.}
\usage{
formatDate(d, m, y, date.format = MM.DD.YYYY, sep = "/")
}
\arguments{
\item{d}{day of the month (1-31)}

\item{m}{month of the year (1-12)}

\item{y}{year of date}

\item{date.format}{how to format the date. Defaults to "month/day/year".}

\item{sep}{string used to separate \code{d}, \code{m}, and \code{y}. Defaults to "/".}
}
\value{
A character string of a formatted date.
}
\description{
Prints a date into a pretty format.
}
\details{
Given the day, month, and year of a date, returns the date in a specific format.
The order and separating string can be modified using \code{date.format} and \code{sep} respectively.
Take note the order of the arguments: day, month, and year. The current version
only accepts "MM.DD.YYYY", "DD.MM.YYYY", "YYYY.MM.DD".
}
\examples{
formatDate(8, 7, 2011)
formatDate(8, 7, 2011, date.format = "YYYY.MM.DD")
formatDate(8, 7, 2011, date.format = "DD.MM.YYYY", sep = "-")
}
\author{
Samuel Leung, Derek Chiu
}

