% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{formatNA}
\alias{formatNA}
\title{Missing Value Formatting}
\usage{
formatNA(y, type = c("cat", "cont"), codes = c("", "Unk", "N/A"))
}
\arguments{
\item{y}{a vector.}

\item{type}{whether the variable is \code{"cat"} (categorical) or
\code{"cont"} (continuous). Defaults to \code{"cat"}.}

\item{codes}{vector of missing codes to replace with \code{NA}}
}
\value{
A categorical or numerical vector with all missing formatted as
\code{NA}.
}
\description{
Takes a numeric vector and replaces all missing codes with NA and returns a
factor if the variable is categorical or a numeric variable if it's numeric.
}
\examples{
y <- c(1:10, "Unk", 12)
formatNA(y)
}
\author{
Aline Talhouk, Derek Chiu
}
