% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqTable.R
\name{freqTable}
\alias{freqTable}
\title{Generate a Frequency Table}
\usage{
freqTable(x, levels, missing = NULL, description = NULL, round = 1,
  plot = FALSE)
}
\arguments{
\item{x}{vector of values to show frequencies for}

\item{levels}{order of levels for \code{x}}

\item{missing}{vector of levels for which we consider missing and don't count
in the valid percentage}

\item{description}{(optional) description for each level of \code{x}.
Must be same length and order as \code{levels}}

\item{round}{number of digits to round percentages to}

\item{plot}{logical; if \code{TRUE}, a barplot is produced.}
}
\value{
A data frame with the following columns
\item{Class}{Tells you which scores are valid and which are missing}
\item{Score}{Different levels}
\item{Frequency}{Count for each score}
\item{Percent}{Percent of Frequency out of the grand total}
\item{Valid Percent}{Percent of Frequency out of the Valid scores}
\item{Cumulative Percent}{Accumulated Percent of Frequency out of the Valid Scores}
\item{Description}{If \code{description} is given, a description for each level}
}
\description{
A frequency table emulating the SPSS FREQ output is generated.
}
\author{
Derek Chiu
}

