% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getFormat.R
\name{getFormat}
\alias{getFormat}
\title{Get date format from character text}
\usage{
getFormat(date, char.format, sep = "")
}
\arguments{
\item{date}{character string of date}

\item{char.format}{character text format of date. Either one of
\code{"MM.DD.YYYY"}, \code{"MMMM.DD.YYYY"}, \code{"DD.MM.YYYY"},
\code{"DD.MMM.YYYY"}, \code{"YYYY.MM.DD"}, or \code{"YYYY.MMM.DD"}.}

\item{sep}{character string separating \code{date}}
}
\value{
A character string representing the POSIX standard date format
equivalent of the string in \code{char.format}.
}
\description{
Get the POSIX standard date formats from character text formats.
}
\examples{
getFormat("12/09/1993", "MM.DD.YYYY")
getFormat("2005-09-13", "YYYY.MM.DD")
}
\author{
Derek Chiu
}

