% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getPval}
\alias{getPval}
\title{Get the p-value}
\usage{
getPval(x)
}
\arguments{
\item{x}{an object from \code{\link[survival:survdiff]{survival::survdiff()}}}
}
\value{
the Chi-squared p-value
}
\description{
Get the p-value
}
\references{
Christos Hatzis
(https://stat.ethz.ch/pipermail/r-help/2007-April/130676.html)
}
