% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggkm.R
\name{ggkm}
\alias{ggkm}
\title{ggkm}
\usage{
ggkm(sfit, sfit2 = NULL, table = TRUE, returns = FALSE, marks = TRUE,
  cols = rainbow(100), shading.colors = c("blue2", "red2", "deepskyblue",
  "indianred3"), xlabs = "Time", ylabs = "Survival Probability",
  xlims = c(0, max(sfit$time)), ylims = c(0, 1), ystratalabs = NULL,
  ystrataname = NULL, timeby = 5, main = "Kaplan-Meier Plot",
  pval = TRUE, HR = TRUE, use.firth = 1, CI = TRUE, subs = NULL,
  legend = FALSE, ...)
}
\description{
ggkm
}

