% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/maxDateArray.R
\name{maxDateArray}
\alias{maxDateArray}
\title{Maximum of date array}
\usage{
maxDateArray(t.arr, date.format = "MM.DD.YYYY", existing.missing.codes = NA,
  return.missing.code = NA)
}
\arguments{
\item{t.arr}{array of date strings}

\item{date.format}{how to format the resulting date}

\item{existing.missing.codes}{missing dates}

\item{return.missing.code}{what to return if there is a missing input}
}
\description{
Finds the maximum (latest) date in an array of date strings.
}

