% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/minDate.R
\name{minDate}
\alias{minDate}
\title{Earliest Date}
\usage{
minDate(dates, na.rm = TRUE)
}
\arguments{
\item{dates}{vector of dates}

\item{na.rm}{logical. Should NAs be removed?}
}
\description{
Earliest Date
}

