% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/minDateArray.R
\name{minDateArray}
\alias{minDateArray}
\title{Minimum date of an array}
\usage{
minDateArray(t.arr, date.format = MM.DD.YYYY, existing.missing.codes = NA,
  return.missing.code = NA)
}
\description{
Minimum date of an array
}

