% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_analyze.R
\name{ms_analyze}
\alias{ms_analyze}
\title{Analyze mass spectrometry data}
\usage{
ms_analyze(x, g, col.names, info.vars = c("Accession", "Sequence",
  "Annotated.Sequence", "Descriptions", "Modifications",
  "Reporter.Quan.Result.ID"))
}
\arguments{
\item{x}{data frame with expression values for both samples}

\item{g}{vector of factor levels for samples}

\item{col.names}{vector of column names for output data frame}

\item{info.vars}{vector of column names containing metadata information. 
These variables are collapsed if not unique.}
}
\value{
A vector of statistics from analyzing mass spec data. Includes
  t-values, Wald p-values, effect sizes, fold change, absolute fold change.
}
\description{
Analyze mass spectrometry data for expression differences between sample 
groups.
}
\author{
Derek Chiu
}

