% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_plot.R
\name{ms_plot}
\alias{ms_plot}
\alias{ms_boxplot}
\alias{ms_plot}
\alias{ms_mean_var}
\title{Boxplot of different MS expression data values}
\usage{
ms_boxplot(x, path, width = 8, height = 10, las = 0)

ms_mean_var(x, g, title, path, width = 8, height = 10, las = 0)
}
\arguments{
\item{x}{data object returned by \code{ms_process}}

\item{path}{path to save boxplot to. Device is pdf.}

\item{width}{width of plot}

\item{height}{height of plot}

\item{las}{angle of axis labels}

\item{g}{vector of treatment groups}

\item{title}{vector of titles for each treatment group}
}
\value{
A pdf with three boxplots is saved to disk location specified by
  \code{path}: the raw data values and the log2 and vsn transformed values.

A pdf is saved to disk location specified by \code{path}: the vsn
  transformed values and mean-sd plots for each treatment group.
}
\description{
Generate boxplot for raw, log2, and vsn data values

Generate mean-sd plots for the VSN data values.
}
\seealso{
Other Mass Spectrometry: \code{\link{ms_analyze}},
  \code{\link{ms_condition}}, \code{\link{ms_process}}

Other Mass Spectrometry: \code{\link{ms_analyze}},
  \code{\link{ms_condition}}, \code{\link{ms_process}}
}
\author{
Derek Chiu

Derek Chiu
}

