% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_plot.R
\name{ms_plot}
\alias{ms_plot}
\alias{ms_boxplot}
\alias{ms_plot}
\alias{ms_mean_var}
\title{Plots for MS analyses}
\usage{
ms_boxplot(x, path, width = 8, height = 10)

ms_mean_var(x, g, title, path, width = 8, height = 10)
}
\arguments{
\item{x}{data object returned by \code{ms_process}}

\item{path}{path to save boxplot to. Device is pdf.}

\item{width}{width of plot}

\item{height}{height of plot}

\item{g}{vector of treatment groups}

\item{title}{vector of titles for each treatment group}
}
\value{
Both functions return a pdf saved to the file location specified by 
  \code{path}. \code{ms_boxplot} shows three boxplots of expression values: 
  raw data values, log2 and vsn transformed values. \code{ms_mean_var} shows 
  the vsn transformed values and mean-sd plots for each treatment group.
}
\description{
\code{ms_boxplot} shows boxplots of different MS expression data values. 
\code{ms_mean_var} shows mean-sd plots for the VSN data values to analyze the
mean-variance relationship.
}
\seealso{
Other Mass Spectrometry: \code{\link{ms_analyze}},
  \code{\link{ms_condition}}, \code{\link{ms_process}}
}
\author{
Derek Chiu
}

