% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_process.R
\name{ms_process}
\alias{ms_process}
\title{Process mass spectrometry data}
\usage{
ms_process(psm, protein, g, sample.id, path, save = TRUE)
}
\arguments{
\item{psm}{PSM file}

\item{protein}{Protein file}

\item{g}{character vector of sample groups}

\item{sample.id}{character vector for sample IDs. Other of samples must match
that in the psm raw data.}

\item{path}{file path to save processed peptide file}

\item{save}{logical; should the peptide file be saved?}
}
\value{
A processed data frame to be used by \code{ms_analyze}.
}
\description{
Process mass spectrometry data for filtering out contaminant samples,
manipulating variables, removing duplicates, and more.
}
\author{
Derek Chiu
}

