% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_top.R
\name{ms_top}
\alias{ms_top}
\title{Top variables in mass spectrometry analysis}
\usage{
ms_top(x, level = c("Gene", "Peptide"), alpha = 0.05, fc = 1.25,
  path = NULL)
}
\arguments{
\item{x}{data frame object returned by \code{ms_summarize}}

\item{level}{analysis is on the "Gene" level or "Peptide" level}

\item{alpha}{significance level}

\item{fc}{absolute fold change}

\item{path}{file path to save result object}
}
\value{
A data frame showing the top variables.
}
\description{
Show top genes/peptides defined by significance level and absolute fold change.
}
\seealso{
Other Mass Spectrometry functions: \code{\link{ms_condition}},
  \code{\link{ms_plot}}, \code{\link{ms_process}},
  \code{\link{ms_summarize}}
}
\author{
Derek Chiu
}

