% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiClassCM.R
\name{multiClassCM}
\alias{multiClassCM}
\title{Confusion matrix summaries}
\usage{
multiClassCM(x, y, seed = 20, num.boot = 1000, conf.level = 0.95,
  digits = 2, method = "wilson")
}
\arguments{
\item{x}{a vector of predicted classes}

\item{y}{a vector of reference classes}

\item{seed}{random seed for bootstrapping}

\item{num.boot}{number of times to bootstrap. Defaults to 1000.}

\item{conf.level}{confidence level. Defaults to 95\%.}

\item{digits}{number of digits to round summaries to}

\item{method}{method for obtaining confidence intervals for binomial 
probabilities. See \code{Hmisc::binconf} for details.}
}
\value{
A confusion matrix for the predicted and reference classes. Then the 
  estimated statistics along with bootstrapped confidence intervals. A
list with the following elements
\item{Accuracy}{Accuracy point estimate, lower bound and upper bound for
bootstrapped CI}
\item{Sensitivity}{Sensitivity point estimate, lower bound and upper bound for
bootstrapped CI}
\item{Specificity}{Specificity point estimate, lower bound and upper bound for
bootstrapped CI}
\item{PPV}{PPV point estimate, lower bound and upper bound for
bootstrapped CI}
\item{NPV}{NPV point estimate, lower bound and upper bound for
bootstrapped CI}
\item{kappa}{kappa point estimate, lower bound and upper bound for
bootstrapped CI}
}
\description{
Calculates summaries from cross-tabulated reference and prediction labels for
a multi-class variable.
}
\details{
Given two multi-class variables summarized in a confusion matrix, this 
function provides performance summaries. It provides overall accuracy with 
confidence intervals, as well as per class accuracy, sensitivity, 
specificity, positive predictive value (PPV), negative predictive value (NPV)
}
\examples{
### 95\% CI from 1000 bootstraped samples
set.seed(547)
n <- 80
x <- rbinom(n, size = 1, prob = 0.6)
y <- rbinom(n, size = 1, prob = 0.4)
binaryCM(x, y)

### 90\% CI from 500 bootstrapped samples
binaryCM(x, y, num.boot = 500, conf.level = 0.90)

### Round to 2 digits
binaryCM(x, y, digits = 2)
}
\author{
Aline Talhouk, Derek Chiu
}
\seealso{
Other confusion matrix functions: \code{\link{binaryCMAsHTML}},
  \code{\link{binaryCM}}
}

