% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_format.R
\name{numericToDate}
\alias{numericToDate}
\title{Change numeric to date}
\usage{
numericToDate(x, date.origin = DATE.ORIGIN)
}
\arguments{
\item{x}{a number that represents the number of days after \code{date.origin}}

\item{date.origin}{the date from which we count the number of days passed}
}
\value{
A date object, converted from a numeric object.
}
\description{
Change a numeric value to a date object by specifying a date of origin.
}
\examples{
numericToDate(10)
numericToDate(10, "2000-09-11")
}
\seealso{
Other date formatting functions: \code{\link{cleanDate}},
  \code{\link{formatDate}}, \code{\link{getFormat}}
}
\author{
Samuel Leung
}
\concept{date formatting functions}
