% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseCor.R
\name{pairwiseCor}
\alias{pairwiseCor}
\title{Pairwise Correlations}
\usage{
pairwiseCor(x)
}
\arguments{
\item{x}{A data frame containing numeric variables of interest.}
}
\value{
all pairwise absolute correlations, correlations, Pval, Adj P val by
decreasing order of absolute correlations.
}
\description{
Computes all pairwise correlations between the columns of a data frame
}
\examples{
set.seed(123)
x <- data.frame(matrix(rnorm(25), nrow = 5))
pairwiseCor(x)
}
\author{
Aline Talhouk, Derek Chiu
}
