% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCAplot.R
\name{PCAplot}
\alias{PCAplot}
\title{Plot Principal Components}
\usage{
PCAplot(by, tdat)
}
\arguments{
\item{by}{factor to plot as ellipses on PCA plots}

\item{tdat}{data matrix to compute principal components}
}
\value{
PCA plots for every combination of PC1, PC2, and PC3. The
percentage of variation contribution is shown in the axes labels.
}
\description{
Plot the first three PCs and draw ellipses highlighting
differences in the levels of a factor
}
\examples{
PCAplot(mtcars$cyl, mtcars)
PCAplot(mtcars$gear, mtcars)
}
\author{
Aline Talhouk, Derek Chiu
}
