% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percents.R
\name{percents}
\alias{colPercent}
\alias{percents}
\alias{rowColPercent}
\alias{rowPercent}
\title{Row and Column Percentages}
\usage{
colPercent(t, pretty.text = FALSE, keep = TRUE, digits = 4)

rowPercent(t, pretty.text = FALSE, keep = TRUE, digits = 4)

rowColPercent(t, keep = TRUE, ...)
}
\arguments{
\item{t}{a matrix}

\item{pretty.text}{logical. If \code{TRUE}, will format the table into nice 
display}

\item{keep}{logical; if \code{TRUE}, the original table counts will be kept 
along with the percentages}

\item{digits}{number of digits to round to}

\item{...}{additional arguments from \code{colPercent} and \code{rowPercent}}
}
\value{
A table with row-wise/column-wise percentages added. The percentages 
  sum to 1 per row/column.
}
\description{
Calculate percentages in a table. \code{rowPercent} gives row percentages, 
\code{colPercent} gives column percentages, and \code{rowColPercent} gives 
both row and column percentages.
}
\details{
Generates a table of row and/or column percentages given table \code{t}. 
Using \code{pretty.text = TRUE} will add the \% sign to the percentages.
}
\examples{
A <- matrix(c(2, 3, 5, 10), nrow = 2, dimnames = list(c("Row1", "Row2"), c("Col1", "Col2")))
rowPercent(A)
rowPercent(A, keep = FALSE)
colPercent(A, pretty.text = TRUE)
colPercent(A, pretty.text = TRUE, keep = FALSE)
rowColPercent(A, digits = 2)
}
\author{
Aline Talhouk, Derek Chiu
}

