% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotKM.R
\name{plotKM}
\alias{plotKM}
\title{Plot Kaplan-Meier curves}
\usage{
plotKM(input.d, input.formula, main.text = "", xlab.text = "",
  ylab.text = "", line.name, line.color, line.pattern = NULL,
  line.width = NULL, show.test = "single",
  single.test.type = "logrank", digits = 3, obs.survyrs = 3,
  legend.pos = "bottomleft", file.name = "no.file", file.width = 7,
  file.height = 7, grey.scale = FALSE, show.single.test.pos = 0.1,
  xlabs = NULL, legend.xy = NULL, ylabs = NULL, timeby = NULL, ...)
}
\arguments{
\item{input.d}{input \code{data.frame}}

\item{input.formula}{survival formula to \code{Surv}}

\item{main.text}{plot title}

\item{xlab.text}{x-axis label}

\item{ylab.text}{y-axis label}

\item{line.name}{name of legend}

\item{line.color}{line colour of survival curves}

\item{line.pattern}{line pattern of survival curves}

\item{line.width}{line width of survival curves}

\item{show.test}{show single or the reference group value (for pairwise
comparisons). If \code{"none"}, then no test is show.}

\item{single.test.type}{test to show if specified \code{show.test = "single"}.
Possible choices are \code{"logrank"} (default), \code{"wilcoxon"}, \code{"taroneware"},
or \code{"all"}.}

\item{digits}{number of digits for p-value}

\item{obs.survyrs}{show the observed survival years survival rate on KM plot}

\item{legend.pos}{legend position keyword}

\item{file.name}{name of file to save plot to}

\item{file.width}{width of figure in saved file}

\item{file.height}{height of figure in saved file}

\item{grey.scale}{logical. If \code{TRUE}, the plot will be in grey scale.}

\item{show.single.test.pos}{position to show single test; defaults to 0.5 if
\code{legend.pos = "top"}. Otherwise 0.1}

\item{xlabs}{dummy param to match ggkm so they don't get passed to ...}

\item{legend.xy}{dummy param to match ggkm so they don't get passed to ...}

\item{ylabs}{dummy param to match ggkm so they don't get passed to ...}

\item{timeby}{dummy param to match ggkm so they don't get passed to ...}

\item{...}{additional arguments to \code{plotKMDetail}}
}
\description{
KM plots with basic annotations.
}
\seealso{
\code{\link[=plotKMDetail]{plotKMDetail()}}
}
\author{
Samuel Leung
}
