% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotKMDetail.R
\name{plotKMDetail}
\alias{plotKMDetail}
\title{Plot detailed Kaplan-Meier curves}
\usage{
plotKMDetail(input.data, surv.formula, main.text = "", xlab.text = "",
  ylab.text = "", line.name, line.color, line.pattern = NULL,
  line.width = NULL, show.test = "single",
  single.test.type = "logrank", round.digits.p.value = 4, obs.survyrs,
  ten.years.surv.95CI, event.count, legend.pos = "bottomleft",
  file.name = "no.file", file.width = 7, file.height = 7,
  grey.scale = FALSE, show.single.test.pos, ...)
}
\arguments{
\item{input.data}{input \code{data.frame}}

\item{surv.formula}{survival formula to \code{Surv}}

\item{main.text}{plot title}

\item{xlab.text}{x-axis label}

\item{ylab.text}{y-axis label}

\item{line.name}{name of legend}

\item{line.color}{line colour of survival curves}

\item{line.pattern}{line pattern of survival curves}

\item{line.width}{line width of survival curves}

\item{show.test}{show single or the reference group value (for pairwise
comparisons). If \code{"none"}, then no test is show.}

\item{single.test.type}{test to show if specified \code{show.test = "single"}. Possible choices are \code{"logrank"} (default),
\code{"wilcoxon"}, \code{"taroneware"}, or \code{"all"}.}

\item{round.digits.p.value}{number of digits for p-value}

\item{obs.survyrs}{show the observed survival years survival rate on KM plot}

\item{ten.years.surv.95CI}{show ten year survival 95\% confidence interval}

\item{event.count}{show the number of events at each time point}

\item{legend.pos}{legend position keyword}

\item{file.name}{name of file to save plot to}

\item{file.width}{width of figure in saved file}

\item{file.height}{height of figure in saved file}

\item{grey.scale}{logical. If \code{TRUE}, the plot will be in grey scale.}

\item{show.single.test.pos}{position to show single test; defaults to 0.5 if
\code{legend.pos = "top"}. Otherwise 0.1}

\item{...}{additional arguments to \code{plot}}
}
\description{
KM plots with details of event counts.
}
\references{
http://courses.nus.edu.sg/course/stacar/internet/st3242/handouts/notes2.pdf
}
\seealso{
\code{\link[=plotKM]{plotKM()}}
}
\author{
Samuel Leung
}
