% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_coxph.R
\name{printCoxMod}
\alias{printCoxMod}
\title{Print Cox model output}
\usage{
printCoxMod(cox, Capt)
}
\arguments{
\item{cox}{is the matrix \code{summary(x)$coef}, where x is an object of
class \code{coxph}}

\item{Capt}{a caption}
}
\description{
Prints a Cox model output in a nice HTML table
}
\author{
Aline Talhouk
}
