% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_small.R
\name{round_small}
\alias{round_small}
\title{Rounding of Small Numbers}
\usage{
round_small(..., digits = 3)
}
\arguments{
\item{...}{a numeric vector}

\item{digits}{integer indicating number of decimal places to round to}
}
\value{
If precision of number is larger than desired rounding, the default
\code{round} is used. Otherwise, we provide an upper bound instead of coercion
to 0.
}
\description{
Rounding of a number smaller than specified precision doesn't coerce to 0.
}
\details{
This function is useful when we have small p-values and don't want to show the
scientific notation, or coercion to 0. Instead we show an upper bound. For example,
if a p-value is 2e-05 and we want to round to 3 digits, the function will return
"<0.001.
}
\examples{
round_small(2e-04)
round_small(5e-04)
round_small(6e-04)
}
\author{
Derek Chiu
}

