% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rowColPercent.R
\name{rowColPercent}
\alias{rowColPercent}
\title{Row and column percentages}
\usage{
rowColPercent(t, ...)
}
\arguments{
\item{t}{a matrix}

\item{...}{additional arguments for \code{\link{rowPercent}}
or \code{\link{colPercent}}}
}
\value{
A table with row and column percentages interleaved.
}
\description{
Calculate row and column percentages in a table.
}
\details{
Details to be filled.
}
\author{
Aline Talhouk, Derek Chiu
}
\seealso{
\code{\link{rowPercent}} for row percentages only,
\code{\link{colPercent}} for column percentages only.
}

