% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rowColPercentAsHTML.R
\name{rowColPercentAsHTML}
\alias{rowColPercentAsHTML}
\title{Row and Column percentages in HTML format}
\usage{
rowColPercentAsHTML(t, show.count = FALSE, row.names = NULL,
  column.names = NULL, html.table.border = 0, banded.rows = FALSE,
  css.class.name.odd = "odd", css.class.name.even = "even", ...)
}
\description{
Row and Column percentages in HTML format
}

