% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rowPercentAsHTML.R
\name{rowPercentAsHTML}
\alias{rowPercentAsHTML}
\title{Row percentages in HTML}
\usage{
rowPercentAsHTML(t, show.count = FALSE, digits = 4, column.names = NULL,
  html.table.border = 0, caption = NA, transpose = FALSE,
  banded.rows = FALSE, css.class.name.odd = "odd",
  css.class.name.even = "even", ...)
}
\description{
Row percentages in HTML
}

