% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sem}
\alias{sem}
\title{Standard error of the mean}
\usage{
sem(x, missing.value = NA, return.missing.value = NA)
}
\arguments{
\item{x}{input vector}

\item{missing.value}{values that are missing}

\item{return.missing.value}{the value to return where there are missing values}
}
\value{
The standard error of the mean of \code{x}
}
\description{
Standard error of the mean
}
\references{
http://en.wikipedia.org/wiki/Standard_error
}
\author{
Samuel Leung
}
