% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startsWith.R
\name{startsWith}
\alias{endsWith}
\alias{startsWith}
\title{Check if a string starts or ends with another string}
\usage{
startsWith(a, b)

endsWith(a, b)
}
\arguments{
\item{a}{string to be checked against}

\item{b}{prefix/suffix to check}
}
\value{
logical; if TRUE, \code{a} is a string that starts or ends with
  \code{b}.
}
\description{
Returns a logical indicating whether the string in the first argument starts 
or ends with the string in the second argument.
}
\details{
Although intended for strings, the function works with numeric vectors as
well.
}
\examples{
# Strings
startsWith("pineapple", "pine")
startsWith("apple", "app")
startsWith("apple", "ppl")
endsWith("pineapple", "apple")
endsWith("apple", "app")

# Numerics
startsWith(123.4, 123)
startsWith(45, 5)
endsWith(123.4, 3.4)
endsWith(45, 4)
}
\author{
Samuel Leung, Derek Chiu
}

