% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryAsHTML.R
\name{summaryAsHTML}
\alias{summaryAsHTML}
\title{Summary table in HTML format}
\usage{
summaryAsHTML(d)
}
\arguments{
\item{d}{assume \code{d} is an array of numbers}
}
\value{
summary table with annotated HTML code
}
\description{
Generate summary table as an HTML table
}
\examples{
library(htmlTable)
set.seed(1)
x <- rnorm(100)
htmlTable(summaryAsHTML(x))
}
\author{
Samuel Leung, Derek Chiu
}
