% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStatsBy.R
\name{SummaryStatsBy}
\alias{SummaryStatsBy}
\title{Generate cohort characteristics}
\usage{
SummaryStatsBy(data, by1, by2, var.names, stats = c("mean", "sd", "median",
  "IQR", "range", "missing"), marker.description = by1,
  groups.description = by2, var.descriptions = var.names, digits = 3,
  format = c("pandoc", "html"))
}
\arguments{
\item{data}{The \code{data.frame} containing the data}

\item{by1}{split variable 1}

\item{by2}{split variable 2}

\item{var.names}{The variables that you want the statistics for}

\item{stats}{statistics to show compute for continuous variables}

\item{marker.description}{The description for the variable(s) to split}

\item{groups.description}{The groupings of the marker}

\item{var.descriptions}{descriptions of \code{var.names}}

\item{digits}{number of digits to round to}

\item{format}{format to return the table in}
}
\description{
Generate cohort characteristics
}
\examples{
mtcars$vs <- as.factor(mtcars$vs); mtcars$am <- as.factor(mtcars$am)
SummaryStatsBy(mtcars, by1 = "cyl", by2 = "gear", var.names = c("mpg", "vs", "qsec", "am"))
SummaryStatsBy(mtcars, by1 = "cyl", by2 = "gear", var.names = c("vs", "qsec"))
}
\author{
Aline Talhouk, Derek Chiu
}

