% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SummaryStatsBy.R
\name{SummaryStatsby}
\alias{SummaryStatsby}
\title{Generate cohort characteristics}
\usage{
SummaryStatsby(input.d, by1, by2, var.names, marker.description = by1,
  groups.description = by2, var.descriptions = var.names, digits = 1,
  missing.codes.highlight = NULL, missing.codes = c("N/A", "", "Unk"),
  decimal = 0, caption = NA)
}
\arguments{
\item{input.d}{The \code{data.frame} containing the data}

\item{var.names}{The variables that you want the statistics for}

\item{marker.description}{The description for the variable(s) to split}

\item{marker.name}{The variable that you want to split into different columns}
}
\description{
Generate cohort characteristics
}
\author{
Aline Talhouk
}

