% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tableAsHTML.R
\name{tableAsHTML}
\alias{tableAsHTML}
\title{Change table to HTML format}
\usage{
tableAsHTML(t, row.names = NULL, column.names = NULL, caption = NA,
  html.table.border = 0, banded.rows = FALSE, css.class.name.odd = "odd",
  css.class.name.even = "even")
}
\description{
Change table to HTML format
}

