% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xunivcoxph.R
\name{Xunivcoxph}
\alias{Xunivcoxph}
\title{Univariate cox proprtional hazards model}
\usage{
Xunivcoxph(mod, coxph.type = c("coxph", "coxphf"), digits = 3)
}
\arguments{
\item{mod}{model fit object, returned from either \code{coxph}
or \code{coxphf}.}

\item{coxph.type}{model fit type; either \code{coxph} (default)
or \code{coxphf} (Firth's correction).}

\item{digits}{number of digits to round}
}
\value{
Hazard ratio, confidence interval, and likelihood ratio test
p-value
}
\description{
Univariate cox proprtional hazards model
}
\author{
Aline Talhouk
}

