% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_coxph.R
\name{Xunivcoxph}
\alias{Xunivcoxph}
\title{Univariate cox proprtional hazards model}
\usage{
Xunivcoxph(mod, digits = 3)
}
\arguments{
\item{mod}{model fit object, returned from either \code{coxph} or
\code{coxphf}.}

\item{digits}{number of digits to round}
}
\value{
Hazard ratio and 95/% confidence interval
}
\description{
Univariate cox proprtional hazards model
}
\author{
Aline Talhouk, Derek Chiu
}

