% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Xunivcoxph.R
\name{Xunivcoxph}
\alias{Xunivcoxph}
\title{Univariate cox proprtional hazards model}
\usage{
Xunivcoxph(mod, digits = 3, coxph.type = "coxph")
}
\description{
Univariate cox proprtional hazards model
}

